fix #3: Output.bmp == Cubic.bmp
