--- _Old_Version/convolution1.h
@@ -1,724 +1,0 @@
-#ifndef __CONVOLUTION1_H_INCLUDED
-#define __CONVOLUTION1_H_INCLUDED
-
-#include <cv.h>
-#include <highgui.h>
-
-#define CONV1_FILTERS	64
-
-void Convolution99(IplImage* src, IplImage* dst, float kernel[9][9], float bias)
-{
-	/*扩展src图像*/
-	IplImage* src2 = cvCreateImage(CvSize(src->width + 8, src->height + 8), IPL_DEPTH_8U, 1);
-	for (int row = 0; row < src2->height; row++)
-	{
-		for (int col = 0; col < src2->width; col++)
-		{
-			int tmpRow = row-4;
-			int tmpCol = col-4;
-
-			if (tmpRow < 0)
-				tmpRow = 0;
-			else if (tmpRow >= src->height)
-				tmpRow = src->height - 1;
-
-			if (tmpCol < 0)
-				tmpCol = 0;
-			else if (tmpCol >= src->width)
-				tmpCol = src->width - 1;
-
-			(src2->imageData + row*src2->widthStep)[col] = (src->imageData + tmpRow*src->widthStep)[tmpCol];
-		}
-	}
-	//cvNamedWindow("Src2");
-	//cvShowImage("Src2", src2);
-
-	/*遍历dst所有像素*/
-	for (int row = 0; row < dst->height; row++)
-	{
-		for (int col = 0; col < dst->width; col++)
-		{
-			/*卷积*/
-			float temp = 0;
-
-			for (int i = 0; i < 9; i++)
-			{
-				for (int j = 0; j < 9; j++)
-				{
-					temp += kernel[i][j] * (unsigned char)(src2->imageData + (row + i)*src2->widthStep)[col + j];
-				}
-			}
-			temp += bias;
-			
-			/*阈值化(取正数，四舍五入)*/
-			unsigned int tempINT = 0;
-			if (temp > 0)
-			{
-				tempINT = (unsigned int)(temp + 0.5);
-			}
-
-			(dst->imageData + row*dst->widthStep)[2 * col + 1] = (unsigned char)tempINT;
-			(dst->imageData + row*dst->widthStep)[2 * col] = (unsigned char)(tempINT >> 8);
-		}
-	}
-
-	return;
-}
-
-//偏置64
-float biases_conv1[CONV1_FILTERS] = {
-	-4.2348, -2.8324, -1.2319, -12.4118, -7.6604, -18.6080, -4.9620, -2.4832,
-	177.2564, -18.7921, -3.4532, 2.6357, -7.5383, 4.7663, -5.5272, 1.0427,
-	-20.4844, -5.3046, -0.0639, 3.9947, -6.2322, -0.0373, -0.0128, -0.3023,
-	-0.0213, 0, -0.0959, 7.5775, -0.0253, -5.1348, -0.0752, 4.9179,
-	1.4223, -17.8580, -6.2216, -0.0045, 1.3004, -2.8590, 1.1567, -6.2761,
-	-5.7466, -0.0414, -0.0443, -0.0040, -16.7272, -1.6781, 0.0017, 6.4334,
-	-0.3501, 7.7397, 0, 0.8378, 1.4096, 5.2955, -2.1651, 6.7147,
-	-11.9314, 2.6340, -6.0306, 7.1022, 13.8739, -20.5740, -2.4089, -11.8912
-};
-
-//权重64*(9*9)
-float weights_conv1_data[CONV1_FILTERS][9][9] = {
-	//01
-	-0.0887, +0.0384, -0.0732, -0.1688, -0.0296, +0.0975, +0.0168, -0.0665, +0.0474,
-	+0.0554, -0.0773, +0.2665, +0.2069, +0.2212, -0.0056, -0.1832, +0.1801, -0.0322,
-	+0.0372, +0.0273, +0.0784, -0.2537, -0.3845, -0.1304, -0.1602, -0.0357, -0.0880,
-	-0.1196, +0.4576, -0.2899, -0.3850, -0.0535, +0.6254, +0.3598, -0.2535, +0.2959,
-	-0.1234, +0.0236, -0.5069, -0.1260, +0.1096, +0.1910, -0.1889, -0.1837, +0.0411,
-	+0.2996, -0.2036, +0.0447, +0.7258, +0.3726, -0.0117, -0.2234, +0.2840, -0.1035,
-	-0.0912, -0.0774, -0.0625, +0.0959, -0.2897, -0.2707, -0.1111, +0.0030, -0.1168,
-	-0.0001, +0.0729, -0.0244, -0.0449, -0.1088, -0.0308, +0.3221, +0.2407, -0.0103,
-	-0.0490, +0.0668, -0.0291, -0.0113, +0.1963, +0.0299, -0.1851, -0.1436, +0.0237,
-	//02
-	-0.0440, -0.0223, -0.0312, +0.1252, -0.0835, +0.1836, -0.0853, +0.2092, -0.0952,
-	+0.0953, -0.0638, -0.0347, +0.1119, -0.3055, +0.1671, -0.2309, +0.0473, -0.1190,
-	+0.1223, +0.0106, +0.0221, +0.3498, -0.1146, +0.0749, -0.2426, +0.1784, -0.1190,
-	-0.0521, -0.2782, -0.2432, +0.2430, -0.1183, +0.1789, -0.0034, +0.2014, +0.1194,
-	+0.0688, -0.2372, -0.0892, +0.2135, -0.3503, +0.1233, -0.0190, -0.2138, +0.0408,
-	+0.0576, -0.2142, -0.0272, +0.3712, -0.0077, +0.1683, +0.0074, -0.1686, -0.0787,
-	+0.0869, -0.0967, -0.0931, +0.1219, -0.0518, -0.0407, -0.0976, +0.0777, +0.1933,
-	-0.1200, +0.1046, -0.0374, -0.1549, -0.1442, +0.1250, +0.0973, +0.1148, -0.0601,
-	+0.0235, +0.0660, +0.0782, +0.0188, -0.0405, +0.1245, -0.1431, +0.0022, -0.0362,
-	//03
-	-0.0351, +0.0577, -0.0555, -0.0107, +0.0548, -0.0759, -0.1210, -0.0253, +0.1356,
-	-0.0213, -0.0607, -0.0075, +0.0884, +0.0740, -0.0129, +0.0095, -0.0097, +0.0333,
-	+0.0447, +0.0237, -0.0018, +0.0683, +0.2151, +0.1007, +0.0460, -0.0010, -0.1286,
-	+0.0351, +0.0408, +0.0063, +0.0210, +0.2578, +0.0737, +0.0671, +0.0433, -0.1312,
-	+0.0477, +0.0195, +0.0833, +0.2621, +0.4272, +0.1595, +0.2190, +0.0881, -0.0092,
-	-0.0360, -0.0471, -0.0353, +0.1194, +0.1822, -0.0861, +0.0269, +0.0182, -0.0194,
-	-0.0982, -0.0026, -0.0587, +0.0421, +0.1253, +0.0067, +0.0053, -0.0128, -0.0059,
-	-0.0746, +0.0131, -0.1171, -0.0641, +0.1039, +0.1076, +0.0007, -0.1080, -0.0551,
-	+0.1633, -0.0484, -0.0497, -0.0479, -0.0105, -0.1022, -0.0428, +0.0446, -0.0394,
-	//04
-	+0.0360, -0.0967, -0.0161, +0.2070, -0.0534, -0.2844, +0.1575, +0.0252, -0.0540,
-	-0.0178, +0.1802, -0.1801, -0.3085, +0.1023, +0.3393, +0.1114, -0.2455, +0.1787,
-	+0.0123, -0.0809, -0.1197, +0.2985, +0.2164, -0.1858, +0.0035, +0.0634, -0.0278,
-	+0.0882, -0.2444, +0.1563, +0.5172, -0.0065, -0.4942, -0.1761, +0.1632, -0.1968,
-	-0.0426, +0.0069, +0.2357, +0.0082, -0.2508, -0.0970, +0.1682, +0.2176, -0.1037,
-	-0.0841, +0.1304, -0.0164, -0.3113, -0.2045, +0.0031, +0.0911, -0.1059, +0.0085,
-	+0.0034, +0.0257, +0.0886, -0.1231, +0.0505, +0.1514, +0.0392, -0.0824, +0.1325,
-	+0.0951, -0.2394, +0.0303, +0.2321, +0.1214, -0.0835, -0.0948, +0.0697, -0.0459,
-	-0.0196, +0.0448, +0.0371, -0.0346, -0.2414, +0.1153, +0.0564, -0.0247, -0.0031,
-	//05
-	-0.0405, +0.0446, +0.0306, +0.0242, -0.1546, +0.2472, -0.0307, +0.0405, -0.0597,
-	+0.0494, -0.0314, -0.1320, +0.2360, -0.0407, -0.0556, -0.3058, +0.2812, -0.0270,
-	-0.1367, +0.0312, -0.0958, -0.0210, -0.1038, +0.2678, -0.0800, -0.0585, -0.1520,
-	+0.1379, +0.1470, +0.0705, +0.1891, -0.2068, -0.0608, +0.1066, +0.0960, +0.1721,
-	-0.1388, -0.4883, -0.1486, +0.5086, -0.0048, -0.3038, -0.0187, -0.0041, -0.1074,
-	+0.2966, -0.4478, +0.1048, +0.6373, -0.0805, -0.1615, +0.0359, +0.1483, -0.0420,
-	+0.1723, -0.3205, -0.1607, +0.2158, -0.2055, +0.1362, -0.0256, +0.0519, -0.0516,
-	-0.1307, +0.2229, -0.1229, +0.0113, -0.1688, +0.2937, -0.1131, -0.0250, +0.0474,
-	-0.0487, +0.1163, -0.1338, +0.1882, -0.1345, +0.0480, -0.1446, +0.1202, -0.0203,
-	//06
-	+0.0792, +0.0073, -0.0081, +0.0343, -0.0991, -0.1717, -0.1265, -0.0479, +0.1224,
-	+0.0199, -0.1222, -0.0051, +0.0978, -0.0075, -0.0431, +0.0248, +0.0134, +0.0392,
-	+0.0484, -0.0532, +0.0212, +0.1344, +0.1985, +0.1439, +0.1059, +0.0348, -0.0693,
-	-0.0025, -0.0576, -0.0080, +0.0520, +0.1476, +0.0383, +0.0989, +0.0772, -0.0154,
-	-0.0304, -0.0203, +0.0881, +0.2355, +0.2309, +0.0874, +0.2456, +0.1009, -0.0161,
-	+0.0430, -0.0242, +0.0141, +0.1829, +0.1373, -0.0292, +0.1253, +0.0447, -0.0458,
-	-0.0205, -0.0055, -0.0259, +0.0969, +0.1309, +0.0640, +0.0616, -0.0204, +0.0135,
-	-0.0661, +0.0100, -0.0773, -0.0504, +0.0929, +0.1276, -0.0241, -0.1516, +0.0241,
-	+0.1210, +0.0240, +0.0329, -0.0483, -0.0527, -0.0280, +0.0072, -0.0035, +0.0169,
-	//07
-	+0.1252, -0.0439, +0.0782, +0.0617, -0.2468, +0.2828, +0.0617, +0.1208, -0.0696,
-	-0.0417, -0.3671, -0.0339, +0.4341, -0.2405, -0.5299, -0.4208, +0.5465, -0.3904,
-	+0.0390, +0.3155, -0.1688, +0.0188, +0.5097, +0.4476, -0.3082, -0.0728, +0.3130,
-	-0.2260, +0.3685, -0.1023, -0.5169, -0.1160, +0.3877, +0.3025, -0.3835, +0.0345,
-	-0.0939, +0.0357, +0.1062, -0.1859, -0.1642, +0.1086, +0.4296, -0.2715, +0.0252,
-	+0.1919, -0.2295, +0.1882, +0.0232, -0.4147, -0.4516, +0.2155, -0.0757, +0.1066,
-	+0.0294, -0.3222, +0.2380, +0.4115, +0.0317, -0.2805, +0.1319, +0.1807, -0.1117,
-	+0.1038, +0.1163, -0.3409, -0.1633, +0.2865, +0.2534, -0.2066, -0.0998, +0.0375,
-	-0.2493, +0.3697, -0.0713, -0.0462, -0.0498, -0.0062, -0.0407, -0.0180, +0.0473,
-	//08
-	+0.0968, +0.1868, +0.0261, +0.0537, +0.1366, -0.0473, -0.2576, -0.0841, +0.1958,
-	-0.0490, -0.0315, -0.0285, -0.0325, +0.0463, +0.0481, -0.0421, -0.0344, -0.1151,
-	-0.0376, +0.0463, +0.0289, -0.0394, +0.0589, +0.0831, +0.0068, +0.0006, -0.0639,
-	+0.1079, +0.1809, +0.0508, -0.0883, +0.0008, +0.0302, +0.0534, +0.1034, +0.0670,
-	-0.0000, +0.1334, +0.0206, -0.1377, -0.0127, +0.0664, +0.0818, +0.0350, +0.0341,
-	-0.1436, -0.0080, -0.0833, -0.1136, -0.0103, +0.0128, +0.0513, -0.0521, +0.0234,
-	-0.1660, -0.0122, -0.0231, +0.0131, +0.0741, +0.0358, +0.1127, -0.0122, -0.0123,
-	-0.0867, +0.0177, +0.0101, +0.0226, +0.0740, +0.0214, +0.0175, -0.0698, -0.0041,
-	+0.1811, -0.1460, -0.1233, +0.1321, +0.1870, +0.0698, +0.0736, +0.0771, +0.0735,
-	//09
-	-0.0506, +0.0104, +0.1018, -0.1497, -0.2431, +0.0038, +0.1177, -0.0077, -0.0575,
-	+0.0854, -0.1154, -0.0427, +0.1641, +0.3261, +0.0294, -0.0907, +0.0572, +0.0387,
-	-0.0671, -0.0027, -0.0524, -0.0600, +0.2165, +0.0520, -0.1874, -0.0743, +0.0429,
-	-0.0185, +0.1099, +0.1457, -0.0156, -0.2260, +0.0074, +0.1142, +0.1068, +0.0043,
-	-0.3400, +0.2609, +0.2239, -0.4687, -0.9168, -0.3209, +0.1911, +0.2351, -0.2881,
-	+0.0596, +0.1558, +0.0180, +0.0082, -0.1766, -0.1594, +0.0438, +0.1101, -0.0126,
-	+0.0211, -0.1030, -0.1259, +0.0112, +0.2135, -0.0867, -0.0197, +0.0053, +0.0295,
-	+0.0218, +0.0172, +0.1043, -0.0006, +0.3220, +0.1743, +0.0346, -0.0925, +0.0663,
-	-0.0675, +0.0387, -0.0332, +0.0005, -0.2200, -0.2017, +0.1065, +0.0834, -0.0432,
-	//10
-	-0.0408, +0.1424, -0.0619, -0.0759, -0.1165, -0.1707, -0.0541, -0.0142, +0.1149,
-	-0.1289, -0.0295, -0.0348, +0.0948, +0.0729, -0.0567, +0.0476, +0.0007, -0.0867,
-	+0.0038, +0.1310, +0.1302, +0.1954, +0.1528, -0.0126, +0.0617, +0.0003, -0.1395,
-	-0.0450, +0.1173, +0.0878, +0.1311, +0.1207, -0.0923, +0.0244, +0.0476, -0.1136,
-	-0.1740, +0.0568, -0.0091, +0.0114, +0.1207, +0.0223, +0.1219, +0.0638, -0.0640,
-	-0.0890, +0.1026, -0.0846, -0.1420, -0.0056, +0.0005, +0.0917, +0.0033, +0.0059,
-	-0.0762, +0.1364, -0.0216, -0.0533, +0.0457, +0.0095, +0.0470, -0.0155, +0.0369,
-	-0.1360, +0.0668, -0.0503, -0.0535, +0.0242, +0.0032, -0.0075, -0.0943, +0.0074,
-	+0.1422, +0.0415, -0.0565, +0.0376, +0.0486, -0.0089, +0.0199, -0.0088, +0.0178,
-	//11
-	+0.1223, -0.1181, +0.0949, +0.0328, -0.0279, +0.0403, -0.1038, +0.0359, +0.0146,
-	-0.1333, -0.0478, +0.0009, -0.1449, +0.0692, +0.1800, -0.1494, +0.0218, -0.0037,
-	+0.0259, +0.1282, +0.1173, -0.0645, -0.0276, +0.1057, -0.1500, +0.0058, +0.0633,
-	+0.1082, -0.0440, -0.1069, +0.0356, -0.0736, +0.0699, +0.0046, +0.0940, -0.0544,
-	+0.1033, -0.2635, -0.4276, +0.1628, +0.0028, -0.0395, -0.0965, +0.0152, -0.0536,
-	+0.2564, +0.0915, -0.2273, +0.4333, +0.1995, -0.0129, -0.0301, +0.0944, +0.0281,
-	-0.0596, +0.1948, -0.2431, -0.0353, -0.3217, -0.1396, +0.0719, +0.0277, -0.0902,
-	-0.1064, +0.2014, -0.0636, +0.0800, -0.2149, +0.0589, +0.1502, -0.0674, +0.0694,
-	+0.0040, -0.0689, -0.0556, +0.3216, -0.0221, -0.0281, +0.0088, -0.0768, +0.0149,
-	//12
-	-0.0052, -0.1051, -0.0017, +0.2346, -0.1947, -0.0382, +0.1557, +0.0438, -0.1761,
-	+0.0138, +0.0554, -0.1016, +0.1225, -0.1940, +0.1057, +0.1794, -0.1317, +0.0654,
-	+0.0624, -0.0266, -0.1965, +0.1610, -0.1389, -0.0559, -0.0745, -0.0400, +0.1083,
-	+0.1019, -0.0162, -0.2107, +0.2875, +0.2884, +0.0480, -0.2856, +0.0389, +0.1202,
-	+0.0531, -0.0673, -0.3648, +0.0658, +0.1756, +0.0371, -0.3539, -0.0975, +0.2053,
-	-0.2313, +0.2951, +0.1471, -0.0500, +0.0589, +0.3200, -0.0236, -0.0521, +0.0229,
-	-0.2259, +0.2490, +0.1371, -0.1670, -0.1568, +0.1098, -0.0337, +0.0107, -0.0052,
-	+0.0343, -0.0459, -0.0302, -0.1130, -0.2033, -0.1539, +0.0102, +0.0531, -0.0416,
-	+0.1307, -0.1318, +0.0417, +0.0323, +0.2553, +0.0288, +0.0146, -0.0034, -0.0175,
-	//13
-	+0.0476, -0.1497, +0.0063, +0.0267, +0.0690, +0.0812, +0.0316, +0.0417, -0.0734,
-	+0.0284, +0.0070, +0.1175, +0.0116, +0.0903, +0.1862, +0.0738, +0.0114, -0.0432,
-	-0.0997, -0.0101, +0.0266, -0.1334, -0.0485, +0.0246, -0.1019, -0.0984, +0.0517,
-	+0.0531, +0.1153, +0.0566, -0.1226, -0.0443, +0.0590, +0.0071, -0.0134, +0.1883,
-	+0.0671, +0.0707, +0.0236, -0.0537, -0.0099, +0.0602, +0.0659, -0.0325, +0.0372,
-	-0.0250, -0.0598, -0.1096, -0.0938, -0.0815, -0.0482, +0.0336, -0.0808, -0.1018,
-	+0.0963, +0.0560, -0.0423, -0.0253, -0.0151, +0.0567, +0.1454, -0.0064, -0.0329,
-	+0.1224, +0.1135, +0.0616, +0.0408, +0.0085, +0.0089, +0.0401, -0.0574, +0.0048,
-	+0.0667, -0.0775, +0.0057, +0.1176, +0.0593, -0.0453, -0.0191, -0.0807, -0.0273,
-	//14
-	+0.1152, -0.0156, -0.1415, +0.0853, +0.0138, +0.0519, -0.0270, +0.0643, -0.0269,
-	-0.1076, +0.0157, -0.0091, +0.1388, -0.0529, +0.0108, -0.0800, +0.0251, +0.0302,
-	-0.0306, +0.1413, -0.0381, -0.0648, -0.1226, +0.0910, -0.0461, -0.0124, -0.0032,
-	-0.0209, +0.1101, +0.0177, -0.0574, -0.0317, +0.2416, +0.0292, -0.0046, -0.0071,
-	-0.0800, +0.0481, +0.1245, -0.0056, -0.1558, +0.1432, -0.0126, -0.1082, -0.0228,
-	-0.1205, +0.0754, +0.1004, -0.0488, -0.3350, -0.0615, -0.1013, -0.2287, -0.1127,
-	-0.0583, +0.0913, -0.0236, +0.0208, -0.1752, +0.0351, -0.0171, -0.0509, +0.0863,
-	-0.1472, +0.0872, -0.0779, +0.1194, +0.0555, +0.2105, +0.1103, +0.0790, +0.1168,
-	-0.0677, +0.2343, -0.0463, +0.0362, -0.0687, -0.0132, -0.0014, +0.0074, -0.1131,
-	//15
-	+0.1639, -0.1876, +0.2133, +0.1075, +0.0256, +0.1558, -0.0599, +0.0446, +0.0034,
-	+0.0831, -0.2639, -0.2249, -0.1600, -0.2004, -0.1886, -0.1489, -0.0482, -0.0567,
-	+0.0441, -0.0045, -0.1273, -0.0501, +0.2019, +0.1388, +0.0946, +0.1883, -0.0882,
-	+0.0804, +0.0693, +0.0915, -0.0870, +0.0930, -0.0195, -0.0851, +0.1910, -0.0644,
-	+0.0637, +0.1245, +0.1290, -0.2069, -0.1128, -0.0979, -0.1397, -0.2810, +0.0617,
-	-0.1023, +0.1422, -0.0307, +0.0712, +0.3926, +0.1824, +0.2480, -0.0956, +0.2063,
-	-0.1028, +0.3288, -0.2737, -0.2348, +0.0753, -0.1631, +0.1179, -0.0247, -0.1313,
-	-0.1145, +0.3420, -0.1041, -0.2623, -0.1266, -0.0884, +0.0507, -0.0895, +0.1396,
-	-0.1320, +0.0818, +0.1472, -0.0135, +0.0676, +0.1195, -0.0121, -0.1381, +0.0287,
-	//16
-	-0.0433, -0.0943, -0.0453, +0.0018, +0.0458, -0.0503, -0.0461, -0.0325, -0.1029,
-	+0.0478, +0.0083, +0.0983, +0.0621, +0.0422, +0.0084, +0.0231, +0.0056, +0.0584,
-	+0.0226, -0.0130, +0.0113, -0.0226, +0.0669, -0.0162, -0.0682, -0.0410, -0.0142,
-	+0.0282, -0.0146, +0.0147, -0.0164, +0.1547, +0.0382, +0.0193, +0.0133, -0.0100,
-	+0.1131, -0.0178, +0.0823, +0.2282, +0.3119, +0.0905, +0.1472, +0.0687, +0.0516,
-	+0.0462, -0.0871, -0.0360, +0.1055, +0.1242, -0.1255, -0.0346, -0.0082, -0.0341,
-	+0.0333, -0.0073, -0.0671, -0.0029, +0.0754, -0.0156, -0.0098, -0.0033, +0.0082,
-	+0.0130, +0.0331, -0.0594, -0.0795, +0.0599, +0.0531, +0.0087, -0.0215, +0.0473,
-	-0.0302, -0.0746, +0.0131, -0.0389, +0.0396, -0.0404, -0.0228, +0.0250, +0.0480,
-	//17
-	+0.0803, -0.2624, +0.1005, -0.0434, +0.1027, -0.1911, +0.1217, -0.0243, +0.0026,
-	-0.0405, +0.1346, +0.3499, -0.1518, +0.0454, +0.0382, +0.0281, -0.2216, +0.1181,
-	+0.0436, -0.1877, -0.0677, +0.0620, +0.2248, -0.0461, +0.0595, +0.1239, -0.0093,
-	+0.0587, +0.1155, +0.0714, +0.0177, -0.1215, -0.4953, -0.1447, +0.2589, -0.2235,
-	-0.1622, +0.1674, +0.1664, -0.0348, +0.0103, -0.0531, +0.0969, +0.1899, -0.0495,
-	-0.0243, +0.1142, -0.1397, -0.4420, -0.0174, +0.1811, +0.0719, -0.2777, +0.0646,
-	+0.0836, +0.0421, +0.0273, -0.1807, +0.0551, +0.0469, +0.1802, -0.0975, +0.0649,
-	+0.0650, -0.2894, +0.0878, +0.3020, +0.1704, -0.2061, -0.0439, +0.0952, -0.0399,
-	-0.0566, +0.1987, -0.0440, -0.1637, -0.0423, +0.0565, -0.0586, +0.0892, -0.0601,
-	//18
-	-0.0022, +0.0619, -0.0894, +0.0200, +0.0754, -0.0289, +0.0975, -0.1995, +0.0512,
-	+0.0757, -0.0139, +0.0123, +0.0676, +0.1172, +0.0470, -0.0559, -0.0246, +0.1691,
-	-0.0735, -0.1507, +0.0706, -0.0719, -0.1910, -0.2694, -0.0827, +0.2131, -0.2611,
-	+0.0651, +0.0346, +0.1422, -0.0124, -0.0148, +0.1877, +0.3172, +0.2718, -0.0741,
-	-0.1237, +0.0477, +0.1115, -0.1049, -0.2297, -0.1411, -0.0658, +0.0165, -0.0291,
-	+0.0143, +0.1371, +0.2414, +0.1758, +0.1163, +0.2597, +0.0942, -0.1394, +0.1346,
-	+0.0321, -0.0940, -0.2094, -0.1299, -0.0190, +0.1068, -0.0716, -0.3443, +0.1153,
-	+0.0983, -0.0633, -0.1661, -0.0470, -0.0419, -0.1853, -0.0697, +0.0016, -0.0354,
-	+0.0304, -0.0844, -0.0156, +0.1526, +0.0608, -0.0190, +0.0266, +0.0039, +0.0565,
-	//19
-	+0.0022, +0.0006, -0.0006, -0.0004, -0.0001, -0.0021, -0.0019, +0.0004, -0.0002,
-	-0.0006, -0.0017, -0.0021, -0.0009, +0.0001, -0.0007, +0.0003, -0.0024, -0.0016,
-	+0.0005, -0.0015, -0.0003, +0.0015, -0.0010, -0.0009, +0.0008, +0.0007, -0.0007,
-	+0.0002, +0.0004, -0.0013, +0.0011, +0.0001, -0.0007, +0.0009, +0.0002, +0.0004,
-	+0.0014, -0.0005, +0.0003, +0.0012, -0.0008, +0.0003, -0.0001, -0.0005, -0.0016,
-	+0.0008, -0.0017, +0.0000, +0.0002, -0.0005, +0.0006, -0.0001, +0.0008, -0.0006,
-	-0.0001, +0.0002, -0.0001, -0.0012, +0.0008, -0.0025, +0.0006, -0.0005, +0.0004,
-	-0.0012, -0.0010, -0.0001, -0.0008, +0.0002, +0.0015, +0.0029, +0.0020, -0.0004,
-	+0.0013, -0.0010, -0.0003, -0.0003, -0.0005, +0.0004, -0.0006, -0.0004, -0.0018,
-	//20
-	-0.0548, +0.0700, -0.1557, -0.1929, -0.1085, +0.0797, +0.0399, -0.1271, +0.0107,
-	+0.2536, +0.0118, +0.1787, +0.4237, +0.2089, -0.0238, -0.0433, +0.2529, +0.0231,
-	-0.2796, -0.1498, -0.1925, -0.1486, -0.1428, +0.1083, -0.2062, -0.0945, -0.0609,
-	+0.1004, +0.3227, +0.1917, -0.0244, -0.3484, +0.1868, +0.0923, +0.0260, +0.0487,
-	-0.1216, -0.4436, -0.0148, +0.2129, -0.5034, -0.2090, +0.0861, +0.0772, -0.2007,
-	+0.2083, -0.1716, +0.2597, +0.6638, +0.3756, +0.0442, -0.2111, +0.1735, +0.0865,
-	+0.1883, -0.1373, -0.0375, +0.0117, +0.1146, +0.2564, -0.0737, +0.0447, +0.0396,
-	+0.0580, -0.3072, -0.2848, -0.2043, -0.4429, -0.1085, +0.0540, -0.0612, -0.1320,
-	-0.1912, +0.3888, +0.0410, +0.0005, +0.0563, +0.1450, -0.0475, -0.0397, +0.0869,
-	//21
-	-0.0972, -0.1382, -0.0318, -0.0178, +0.0850, -0.0008, +0.0242, +0.0740, +0.0204,
-	+0.0321, +0.0544, +0.1499, +0.0483, +0.0398, +0.0293, +0.0927, +0.0950, +0.1741,
-	-0.0756, -0.0281, +0.0189, -0.0649, +0.0049, -0.0730, -0.0748, -0.0686, -0.0039,
-	-0.0124, -0.0196, +0.0475, -0.0168, +0.1605, +0.0316, -0.0161, -0.0899, -0.0747,
-	+0.0969, -0.0428, +0.0583, +0.1504, +0.3090, +0.0960, +0.0766, -0.0306, +0.0396,
-	-0.0742, -0.0952, -0.0398, +0.0083, +0.1315, -0.0707, -0.0560, -0.0619, -0.0305,
-	-0.0239, +0.0650, -0.0105, -0.0502, +0.0505, -0.0191, -0.0070, +0.0022, -0.0096,
-	+0.0455, +0.1578, +0.0388, -0.0474, +0.0186, -0.0222, +0.0568, +0.0816, +0.0186,
-	-0.0826, -0.0681, +0.0411, +0.0132, +0.0633, -0.0989, -0.0182, +0.0517, -0.0447,
-	//22
-	+0.0011, -0.0009, +0.0005, +0.0007, +0.0026, -0.0007, -0.0002, -0.0002, +0.0014,
-	-0.0011, -0.0007, +0.0014, -0.0022, -0.0006, -0.0018, -0.0016, -0.0002, +0.0017,
-	+0.0003, -0.0005, -0.0016, +0.0006, -0.0006, +0.0002, -0.0007, +0.0005, -0.0011,
-	+0.0005, +0.0013, +0.0007, -0.0009, +0.0014, +0.0009, +0.0014, +0.0013, -0.0015,
-	+0.0001, -0.0004, -0.0014, -0.0003, +0.0015, +0.0002, -0.0009, -0.0010, -0.0006,
-	-0.0012, +0.0019, -0.0001, -0.0010, -0.0014, +0.0002, -0.0023, +0.0010, -0.0007,
-	+0.0002, +0.0009, +0.0007, +0.0010, +0.0000, +0.0020, -0.0012, +0.0005, -0.0008,
-	+0.0006, -0.0006, -0.0009, -0.0015, +0.0005, -0.0006, -0.0015, -0.0006, -0.0009,
-	+0.0008, -0.0006, -0.0024, -0.0001, -0.0004, -0.0011, +0.0002, +0.0007, -0.0006,
-	//23
-	+0.0020, -0.0010, +0.0003, -0.0006, -0.0013, +0.0021, -0.0001, +0.0009, -0.0017,
-	+0.0004, -0.0008, -0.0002, +0.0021, -0.0002, -0.0007, -0.0001, -0.0014, -0.0004,
-	-0.0010, -0.0022, -0.0019, +0.0004, -0.0001, -0.0012, -0.0000, -0.0002, -0.0000,
-	+0.0011, +0.0006, -0.0002, -0.0003, +0.0004, +0.0005, +0.0008, +0.0008, -0.0002,
-	+0.0008, -0.0008, +0.0002, -0.0003, -0.0003, +0.0012, -0.0007, -0.0008, +0.0003,
-	+0.0005, +0.0001, -0.0008, -0.0004, -0.0004, +0.0009, +0.0010, -0.0001, -0.0008,
-	+0.0000, -0.0007, -0.0013, +0.0001, -0.0020, +0.0004, +0.0001, -0.0007, +0.0013,
-	-0.0005, +0.0000, +0.0001, -0.0014, +0.0013, -0.0001, -0.0014, -0.0013, +0.0001,
-	-0.0005, +0.0007, +0.0019, -0.0009, -0.0021, -0.0008, +0.0006, -0.0002, +0.0015,
-	//24
-	-0.0010, +0.0006, -0.0022, -0.0006, -0.0023, -0.0030, -0.0000, -0.0034, -0.0032,
-	+0.0007, +0.0016, -0.0028, -0.0003, -0.0007, -0.0009, +0.0005, -0.0010, -0.0023,
-	+0.0030, +0.0034, +0.0015, +0.0010, -0.0000, -0.0000, +0.0007, +0.0004, +0.0002,
-	+0.0010, +0.0012, +0.0020, +0.0023, +0.0010, +0.0007, -0.0013, +0.0020, +0.0003,
-	+0.0012, +0.0024, +0.0009, +0.0011, +0.0009, +0.0006, +0.0000, -0.0007, -0.0005,
-	-0.0031, -0.0013, -0.0001, +0.0008, +0.0011, -0.0000, -0.0004, -0.0006, +0.0011,
-	-0.0017, +0.0004, -0.0032, -0.0016, -0.0001, -0.0010, -0.0004, +0.0000, +0.0002,
-	+0.0002, -0.0030, -0.0004, -0.0027, -0.0010, +0.0002, -0.0008, +0.0005, +0.0007,
-	-0.0017, -0.0021, -0.0021, -0.0015, -0.0010, -0.0002, -0.0005, +0.0019, +0.0005,
-	//25
-	+0.0009, +0.0001, +0.0002, -0.0013, -0.0003, -0.0027, +0.0024, +0.0012, -0.0016,
-	+0.0007, -0.0005, -0.0012, -0.0011, +0.0019, +0.0011, +0.0001, +0.0002, +0.0003,
-	-0.0003, -0.0000, -0.0015, -0.0010, +0.0008, -0.0001, +0.0005, +0.0005, -0.0000,
-	-0.0013, +0.0005, +0.0001, -0.0022, -0.0006, +0.0005, +0.0011, +0.0009, -0.0008,
-	+0.0004, -0.0002, +0.0006, +0.0008, -0.0008, -0.0006, +0.0001, +0.0009, -0.0001,
-	-0.0001, -0.0005, -0.0006, -0.0017, -0.0003, -0.0007, +0.0009, -0.0002, +0.0004,
-	+0.0012, +0.0006, +0.0004, -0.0000, -0.0010, -0.0007, +0.0012, +0.0003, -0.0003,
-	-0.0003, -0.0009, -0.0014, +0.0003, +0.0004, -0.0025, -0.0012, -0.0019, -0.0010,
-	+0.0004, -0.0014, -0.0001, -0.0010, -0.0004, +0.0005, +0.0005, +0.0005, -0.0014,
-	//26
-	+0.0004, +0.0002, +0.0017, -0.0005, -0.0018, +0.0013, -0.0011, -0.0009, -0.0004,
-	-0.0012, +0.0001, -0.0013, -0.0005, +0.0004, +0.0006, +0.0004, -0.0018, +0.0010,
-	-0.0009, +0.0014, +0.0005, -0.0006, -0.0009, +0.0010, +0.0000, -0.0012, -0.0002,
-	+0.0000, +0.0000, -0.0004, +0.0013, -0.0021, -0.0011, +0.0000, -0.0021, -0.0002,
-	-0.0019, +0.0012, -0.0009, +0.0021, +0.0001, +0.0007, +0.0018, -0.0003, +0.0003,
-	-0.0002, +0.0008, +0.0002, -0.0011, +0.0005, -0.0012, -0.0015, +0.0003, +0.0004,
-	-0.0019, -0.0020, -0.0014, -0.0005, -0.0010, -0.0010, +0.0007, +0.0003, -0.0034,
-	+0.0018, -0.0001, -0.0008, +0.0023, -0.0018, -0.0003, +0.0001, -0.0007, -0.0018,
-	+0.0002, -0.0006, +0.0016, -0.0003, -0.0021, -0.0001, -0.0007, -0.0002, +0.0000,
-	//27
-	+0.0006, -0.0008, -0.0005, +0.0009, +0.0007, +0.0001, +0.0002, +0.0009, +0.0010,
-	-0.0010, +0.0001, -0.0003, +0.0000, +0.0002, +0.0010, -0.0012, -0.0000, -0.0005,
-	-0.0000, +0.0001, +0.0015, +0.0016, -0.0007, -0.0005, -0.0004, +0.0010, +0.0002,
-	-0.0017, -0.0012, -0.0006, -0.0018, -0.0008, +0.0003, -0.0008, -0.0006, -0.0008,
-	-0.0024, +0.0026, -0.0012, -0.0015, +0.0002, -0.0012, -0.0003, +0.0008, -0.0000,
-	-0.0001, -0.0004, -0.0008, +0.0007, -0.0007, +0.0025, +0.0007, -0.0002, -0.0033,
-	+0.0006, -0.0007, -0.0015, +0.0017, -0.0016, +0.0002, -0.0004, +0.0011, -0.0011,
-	-0.0009, -0.0011, -0.0019, -0.0001, -0.0001, -0.0007, +0.0002, +0.0002, -0.0001,
-	-0.0004, +0.0006, +0.0006, -0.0001, +0.0003, -0.0008, -0.0014, +0.0004, -0.0014,
-	//28
-	-0.0962, +0.0135, +0.1383, -0.0052, -0.0881, -0.1419, +0.3318, -0.2889, +0.1391,
-	+0.1048, -0.0307, -0.0988, -0.0710, +0.2232, +0.0414, -0.0507, -0.0542, +0.0153,
-	-0.0632, +0.0267, -0.0405, -0.0627, +0.1700, -0.0495, -0.1829, +0.1738, -0.0404,
-	+0.0398, -0.1088, +0.2181, +0.1772, -0.2426, -0.3672, +0.0852, +0.3603, -0.1701,
-	-0.0014, +0.1240, -0.0090, -0.0671, -0.2325, +0.0673, +0.2672, -0.1871, -0.1220,
-	-0.1515, +0.2411, -0.2286, -0.3108, +0.1449, +0.6864, +0.1154, -0.3229, +0.2216,
-	+0.0397, -0.0677, -0.0201, +0.1039, -0.0112, +0.4052, -0.0686, -0.1750, +0.0035,
-	+0.0531, -0.0677, +0.1480, +0.4285, -0.3302, -0.2975, -0.2457, -0.1973, -0.0568,
-	-0.0744, +0.1524, -0.2521, -0.0100, -0.0906, +0.1074, +0.1606, +0.0199, +0.1398,
-	//29
-	-0.0010, -0.0011, -0.0010, +0.0001, +0.0007, -0.0012, +0.0006, +0.0007, -0.0007,
-	-0.0018, +0.0001, +0.0006, +0.0020, +0.0001, +0.0013, -0.0005, -0.0004, -0.0015,
-	-0.0000, -0.0006, -0.0004, -0.0013, -0.0008, -0.0001, -0.0009, +0.0003, +0.0001,
-	+0.0010, -0.0004, +0.0007, +0.0017, +0.0008, -0.0005, -0.0004, -0.0002, +0.0002,
-	+0.0023, +0.0001, +0.0007, -0.0010, +0.0010, +0.0001, -0.0005, +0.0005, -0.0017,
-	+0.0005, -0.0002, -0.0002, -0.0012, +0.0014, -0.0007, -0.0003, -0.0004, +0.0007,
-	-0.0008, -0.0008, -0.0020, -0.0011, -0.0018, +0.0007, +0.0002, +0.0003, +0.0011,
-	-0.0000, -0.0012, +0.0008, +0.0007, -0.0008, -0.0007, -0.0008, +0.0002, -0.0006,
-	+0.0005, -0.0006, +0.0002, -0.0009, +0.0001, -0.0003, +0.0007, +0.0004, -0.0010,
-	//30
-	-0.0013, +0.0069, +0.0211, -0.0691, +0.1102, -0.0433, -0.0463, +0.0435, -0.0609,
-	+0.0168, -0.0773, +0.1138, -0.0674, +0.0206, +0.0534, +0.0435, -0.0598, +0.0622,
-	+0.2630, -0.2436, +0.1282, +0.0151, -0.1716, -0.1066, +0.0632, -0.1074, -0.0708,
-	+0.1398, -0.3049, -0.1203, +0.0608, +0.1323, -0.0842, +0.0147, +0.1163, +0.1428,
-	-0.1139, +0.0015, -0.0821, -0.0257, +0.3949, -0.0134, -0.0750, +0.0206, +0.0168,
-	-0.1066, +0.1969, +0.1863, -0.0929, +0.2267, -0.0803, -0.1232, -0.1612, +0.0082,
-	-0.0652, -0.1137, +0.0803, -0.1736, -0.1319, -0.2836, -0.0064, +0.0176, +0.1561,
-	+0.0773, -0.2005, +0.0398, -0.0443, +0.0743, -0.0710, +0.2899, +0.2959, +0.0140,
-	+0.0179, +0.0550, +0.1026, -0.1304, +0.2273, -0.0238, -0.1076, -0.0746, -0.1754,
-	//31
-	-0.0015, +0.0014, -0.0012, -0.0030, -0.0024, +0.0006, -0.0017, -0.0013, -0.0006,
-	+0.0003, +0.0000, +0.0011, +0.0021, +0.0004, +0.0007, +0.0012, -0.0010, -0.0008,
-	-0.0007, +0.0001, -0.0006, +0.0007, -0.0018, +0.0001, -0.0003, +0.0003, +0.0000,
-	+0.0004, +0.0003, -0.0021, +0.0018, +0.0010, -0.0001, -0.0015, +0.0016, -0.0000,
-	-0.0001, -0.0003, -0.0003, -0.0018, -0.0001, +0.0011, -0.0005, +0.0008, +0.0004,
-	-0.0000, -0.0020, +0.0011, +0.0002, -0.0013, +0.0016, -0.0007, -0.0009, -0.0009,
-	+0.0011, -0.0001, +0.0003, +0.0006, +0.0003, +0.0002, -0.0007, -0.0003, -0.0005,
-	-0.0003, -0.0012, -0.0012, -0.0012, -0.0004, +0.0020, +0.0001, +0.0002, -0.0008,
-	-0.0018, +0.0008, +0.0003, -0.0009, +0.0001, -0.0016, -0.0002, -0.0014, +0.0001,
-	//32
-	-0.1545, -0.0855, +0.0621, +0.2066, -0.0607, -0.0230, -0.0403, +0.0343, -0.0445,
-	+0.1910, -0.1137, +0.1787, +0.0300, -0.2106, +0.1374, +0.0909, -0.0836, -0.0183,
-	+0.2047, -0.4546, -0.1051, +0.2124, -0.0444, -0.1268, -0.1643, +0.1697, +0.1134,
-	+0.0261, -0.1696, -0.0650, +0.3269, +0.0901, +0.2075, -0.1691, -0.0670, -0.1108,
-	+0.0902, +0.0817, -0.2464, -0.1386, -0.2218, +0.5987, +0.0767, -0.1749, +0.0917,
-	-0.0387, +0.3894, -0.2261, -0.2789, -0.0703, +0.4472, -0.1565, -0.2485, +0.1784,
-	-0.2947, +0.3428, +0.1996, -0.3092, -0.0516, +0.1274, -0.1333, +0.0165, -0.0385,
-	+0.1117, -0.3309, +0.2003, -0.0003, +0.1167, -0.0619, -0.0163, +0.1311, +0.0244,
-	+0.1267, -0.1779, +0.0910, -0.1499, +0.0445, -0.0699, +0.0471, +0.0290, -0.0285,
-	//33
-	+0.0649, +0.0384, -0.0558, -0.0471, -0.0204, +0.0108, +0.0922, +0.1087, -0.0027,
-	+0.0302, +0.0083, -0.0084, +0.0363, +0.0033, -0.0081, +0.0567, +0.0298, -0.0515,
-	+0.0666, +0.0571, +0.0088, +0.0207, +0.0148, -0.0218, +0.0332, +0.0479, +0.0021,
-	+0.0156, +0.0218, -0.0130, -0.0064, +0.0243, -0.0523, +0.0097, +0.0770, +0.0487,
-	-0.0084, +0.0210, +0.0306, +0.0940, +0.1340, +0.0176, +0.0346, +0.0494, +0.0505,
-	-0.0102, +0.0398, +0.0199, +0.0365, +0.0479, -0.0315, -0.0210, +0.0044, +0.0562,
-	-0.0454, +0.0191, -0.0145, -0.0167, +0.0053, +0.0052, -0.0029, +0.0270, +0.1252,
-	-0.0711, -0.0167, -0.0545, -0.0663, -0.0056, +0.0456, +0.0024, -0.0208, +0.0856,
-	+0.0089, -0.0001, +0.0083, -0.0218, -0.0114, -0.0257, -0.0504, -0.0314, +0.0100,
-	//34
-	+0.0488, -0.0011, -0.0058, -0.2583, +0.0281, +0.0841, +0.1278, -0.1893, +0.1209,
-	-0.0401, -0.1240, +0.3117, +0.0781, +0.1313, -0.2873, +0.0404, -0.0119, -0.0785,
-	-0.0249, -0.0796, +0.1779, -0.2475, +0.2987, +0.1337, +0.0541, -0.1553, +0.1198,
-	+0.1093, -0.0857, -0.1938, -0.4628, -0.1732, -0.0451, +0.2324, -0.1144, +0.0348,
-	+0.0264, +0.2911, +0.1001, +0.0906, +0.0654, -0.1610, +0.1837, +0.1240, -0.0482,
-	-0.1824, +0.0784, +0.0926, +0.0565, -0.0298, -0.4304, -0.2501, +0.1574, -0.1013,
-	+0.0696, -0.1580, +0.2208, +0.2824, +0.1145, -0.0829, -0.0556, -0.0078, -0.1230,
-	+0.0461, -0.1864, -0.1330, -0.2103, +0.1277, +0.3030, +0.1179, +0.0670, +0.0524,
-	+0.0630, +0.0088, +0.0215, +0.0999, +0.0693, -0.2347, -0.1826, +0.1828, -0.0557,
-	//35
-	-0.1677, +0.2868, -0.0938, +0.0451, -0.2225, +0.2253, -0.0567, -0.1102, +0.0702,
-	+0.1111, -0.0626, -0.3951, +0.3311, +0.0124, +0.1720, -0.2461, +0.1565, +0.0115,
-	-0.0107, +0.1896, -0.0455, +0.2732, -0.2368, +0.1033, -0.2571, -0.1229, -0.0142,
-	-0.1919, +0.1186, -0.0525, -0.2056, -0.3588, +0.4819, +0.1379, -0.0796, +0.1059,
-	-0.0564, +0.0548, +0.0795, +0.0079, -0.0518, +0.6536, -0.1367, -0.3603, -0.1155,
-	+0.0571, +0.1320, -0.0981, +0.0269, -0.2892, +0.1393, -0.0291, +0.0441, +0.0718,
-	-0.1125, +0.0644, -0.1416, +0.2121, -0.1129, -0.1281, -0.0098, +0.2383, -0.1557,
-	-0.0541, +0.1291, -0.1262, +0.0067, +0.1451, +0.2919, -0.2008, -0.0463, -0.0151,
-	+0.0396, -0.0335, +0.0303, +0.0735, -0.2257, -0.0115, -0.0156, +0.1948, -0.0775,
-	//36
-	-0.0011, -0.0001, -0.0005, -0.0011, -0.0001, -0.0011, +0.0007, -0.0009, +0.0011,
-	-0.0001, -0.0002, -0.0003, -0.0004, -0.0008, -0.0003, +0.0016, +0.0018, +0.0005,
-	-0.0012, -0.0012, -0.0002, -0.0002, -0.0002, +0.0004, -0.0003, +0.0006, -0.0018,
-	+0.0011, -0.0017, -0.0019, +0.0020, -0.0005, +0.0014, -0.0001, -0.0004, -0.0014,
-	-0.0004, +0.0009, +0.0022, +0.0007, +0.0001, -0.0006, -0.0005, -0.0009, -0.0003,
-	+0.0007, -0.0007, -0.0017, +0.0019, +0.0005, -0.0008, -0.0000, -0.0002, -0.0006,
-	+0.0007, -0.0026, +0.0001, -0.0014, +0.0010, +0.0011, -0.0007, +0.0011, -0.0014,
-	-0.0005, -0.0004, +0.0001, -0.0002, -0.0004, +0.0006, -0.0008, -0.0012, -0.0000,
-	+0.0008, -0.0007, -0.0017, -0.0011, +0.0011, +0.0003, -0.0006, -0.0007, -0.0000,
-	//36
-	-0.2105, +0.1674, -0.1903, +0.0022, +0.1397, +0.0103, +0.0350, -0.0418, -0.0696,
-	+0.1213, +0.0539, +0.1327, +0.0807, -0.1188, -0.0074, +0.1879, +0.0871, +0.0866,
-	+0.0294, -0.0144, +0.0709, +0.0570, -0.1823, -0.0724, -0.0479, +0.0204, +0.0837,
-	-0.1171, +0.0514, -0.1285, -0.0756, +0.0229, +0.1051, -0.3395, -0.0475, +0.1691,
-	-0.0278, +0.0875, -0.1150, +0.0156, +0.0950, +0.1838, -0.2302, -0.0730, +0.0505,
-	-0.0542, -0.0296, -0.1456, +0.1535, -0.0328, -0.0484, -0.0285, -0.0362, -0.1243,
-	+0.0390, -0.0081, -0.0628, +0.0963, -0.2108, -0.1806, +0.0546, +0.0976, -0.0210,
-	+0.1672, -0.0426, +0.0377, +0.1446, -0.0021, +0.1259, +0.0033, +0.0751, +0.0500,
-	+0.1155, -0.2443, -0.0305, +0.0984, -0.0381, +0.1052, -0.1479, +0.1318, -0.0955,
-	//37
-	+0.0981, +0.0052, +0.0846, +0.1987, +0.2173, -0.0286, +0.0186, -0.0906, +0.0777,
-	-0.0422, -0.4452, -0.0952, -0.2061, -0.3937, -0.5098, +0.0004, +0.1993, -0.3002,
-	+0.1126, +0.0786, +0.0294, -0.1686, +0.1669, +0.2001, +0.2500, +0.1679, +0.0911,
-	-0.0979, +0.3587, +0.0405, +0.0889, +0.4914, -0.0140, -0.2397, -0.1402, -0.0585,
-	-0.0094, -0.0033, -0.2303, +0.0444, +0.2518, -0.2463, -0.0486, +0.1816, +0.1030,
-	+0.1038, +0.1490, -0.1145, -0.3011, +0.0205, -0.0366, +0.0634, -0.1683, -0.0233,
-	-0.3148, +0.2253, +0.1440, -0.4742, -0.0641, +0.0439, +0.1972, -0.1834, +0.0855,
-	+0.0417, +0.0661, +0.2597, +0.1445, +0.2552, -0.2458, +0.1495, +0.0042, -0.0665,
-	-0.0075, -0.0053, -0.0693, -0.2126, +0.1025, -0.0462, +0.0579, -0.1690, +0.1569,
-	//39
-	-0.0140, -0.0664, +0.1002, -0.1322, +0.0160, -0.1467, +0.1364, -0.1965, +0.1736,
-	+0.1401, -0.0201, +0.1323, +0.0195, +0.2026, -0.3198, +0.0233, +0.2189, -0.1102,
-	+0.1036, -0.4639, +0.1629, +0.1850, +0.2550, -0.4732, -0.1069, +0.3164, -0.2651,
-	+0.0575, -0.3032, +0.2756, -0.2929, +0.1082, -0.1772, +0.2146, +0.4430, -0.1763,
-	-0.0354, +0.1479, +0.2846, -0.5955, +0.2742, +0.2180, +0.0375, -0.0658, -0.0874,
-	-0.1468, +0.1474, +0.2592, -0.6025, -0.0038, +0.1921, -0.1499, -0.2093, +0.2281,
-	-0.0029, +0.1099, +0.1239, -0.1888, +0.1015, +0.2134, -0.0436, -0.0672, +0.1124,
-	-0.0388, +0.0549, -0.2248, -0.1020, +0.1486, +0.0030, -0.1408, -0.0111, +0.0116,
-	+0.0525, -0.1166, +0.1179, +0.0267, -0.0956, +0.1296, +0.0898, -0.0605, -0.0370,
-	//40
-	-0.0805, +0.1040, -0.0760, -0.1633, -0.0852, -0.0357, -0.2193, -0.3259, +0.0457,
-	+0.0341, +0.0534, +0.1764, +0.0848, +0.4950, +0.7016, +0.3555, +0.0642, +0.3343,
-	+0.0202, -0.0667, +0.0246, -0.1735, -0.0341, +0.2024, +0.0595, -0.1176, -0.1148,
-	+0.0325, +0.0212, -0.0612, -0.2237, -0.4058, -0.3879, -0.1995, +0.0570, -0.1974,
-	-0.0568, +0.3165, +0.1500, -0.0292, -0.0927, +0.1195, +0.1067, +0.0677, +0.1505,
-	-0.2276, -0.0198, -0.1698, -0.1097, -0.1825, -0.1603, -0.0234, -0.0488, -0.1380,
-	+0.0140, -0.0976, -0.0239, +0.1252, +0.1732, -0.1442, +0.0096, +0.1707, -0.0721,
-	+0.1582, +0.0752, +0.1034, -0.1127, +0.1315, +0.1156, +0.0429, -0.0104, +0.0300,
-	-0.0947, -0.1059, +0.0567, +0.1027, -0.0039, -0.2670, -0.0579, +0.1252, -0.0371,
-	//41
-	+0.0394, +0.2202, -0.1815, -0.1696, +0.0960, +0.0997, +0.1203, -0.1693, +0.1674,
-	-0.2363, +0.0672, +0.0455, -0.1155, +0.0736, +0.0304, -0.0252, -0.2344, -0.0686,
-	+0.0335, +0.0159, +0.0661, -0.1932, +0.0232, +0.1044, +0.1430, -0.0179, -0.0307,
-	-0.0140, +0.1144, +0.2554, -0.2266, -0.1133, -0.0363, +0.0564, -0.0065, +0.0716,
-	-0.0505, +0.0321, +0.1270, -0.1401, +0.0763, -0.0244, +0.0295, -0.0344, -0.0056,
-	-0.0151, -0.0171, +0.0785, -0.0434, +0.0934, -0.1236, +0.1478, +0.1657, -0.0935,
-	-0.0270, -0.0106, +0.1942, +0.0483, +0.0579, -0.1842, +0.0716, +0.0789, -0.0686,
-	+0.0143, -0.2818, -0.0119, -0.0202, +0.1115, -0.1510, +0.0563, +0.0999, +0.0702,
-	+0.0802, -0.1185, +0.1792, -0.0256, +0.1032, -0.2012, -0.0183, -0.0116, -0.0793,
-	//42
-	-0.0006, +0.0008, +0.0002, -0.0019, +0.0000, -0.0002, -0.0007, +0.0009, -0.0011,
-	+0.0005, -0.0014, +0.0011, -0.0008, -0.0016, -0.0004, -0.0018, +0.0008, +0.0015,
-	+0.0004, -0.0006, -0.0007, +0.0008, +0.0007, -0.0008, -0.0003, +0.0008, -0.0020,
-	+0.0008, -0.0008, +0.0001, +0.0019, +0.0002, +0.0001, +0.0001, -0.0014, +0.0002,
-	+0.0007, -0.0006, -0.0003, -0.0011, -0.0003, +0.0014, -0.0015, +0.0007, -0.0010,
-	+0.0004, -0.0004, +0.0007, +0.0005, -0.0004, -0.0007, -0.0001, +0.0019, +0.0003,
-	-0.0009, -0.0003, +0.0006, +0.0018, -0.0010, -0.0012, +0.0012, +0.0009, -0.0004,
-	+0.0002, -0.0003, +0.0005, -0.0024, -0.0007, +0.0001, +0.0003, -0.0007, -0.0005,
-	-0.0007, +0.0014, -0.0008, +0.0003, +0.0008, -0.0000, +0.0006, +0.0008, -0.0016,
-	//43
-	-0.0004, -0.0020, +0.0004, -0.0004, +0.0006, -0.0009, +0.0019, +0.0006, -0.0004,
-	+0.0011, +0.0011, -0.0012, -0.0019, +0.0004, +0.0006, -0.0006, +0.0007, +0.0016,
-	+0.0006, -0.0002, -0.0014, +0.0009, -0.0012, +0.0002, -0.0002, -0.0005, -0.0014,
-	-0.0004, +0.0008, -0.0016, +0.0012, +0.0008, -0.0014, -0.0004, -0.0007, -0.0003,
-	-0.0004, +0.0010, -0.0028, +0.0012, -0.0001, -0.0014, +0.0005, -0.0008, -0.0016,
-	+0.0004, +0.0005, -0.0009, -0.0013, -0.0009, +0.0006, +0.0018, -0.0000, -0.0008,
-	+0.0006, -0.0009, -0.0022, -0.0009, +0.0005, +0.0025, +0.0004, +0.0003, -0.0015,
-	+0.0007, -0.0009, +0.0005, +0.0006, -0.0012, -0.0000, -0.0007, +0.0014, +0.0006,
-	-0.0005, +0.0005, -0.0006, -0.0004, -0.0000, +0.0008, -0.0004, -0.0003, -0.0017,
-	//44
-	-0.0004, -0.0003, -0.0002, +0.0006, -0.0001, +0.0010, +0.0005, +0.0001, -0.0000,
-	-0.0010, +0.0015, -0.0014, +0.0010, -0.0012, +0.0003, -0.0007, -0.0011, -0.0002,
-	-0.0011, -0.0008, +0.0003, -0.0002, -0.0000, +0.0004, -0.0008, +0.0004, +0.0007,
-	+0.0018, +0.0010, +0.0010, -0.0004, +0.0012, +0.0010, -0.0006, +0.0005, -0.0006,
-	-0.0012, -0.0003, -0.0010, +0.0010, -0.0009, -0.0001, +0.0003, -0.0009, +0.0007,
-	-0.0003, +0.0003, +0.0010, -0.0012, -0.0016, -0.0009, +0.0002, -0.0001, -0.0002,
-	-0.0005, +0.0004, -0.0007, -0.0009, -0.0007, -0.0007, +0.0008, -0.0000, +0.0000,
-	+0.0005, -0.0020, -0.0013, -0.0002, +0.0001, +0.0001, -0.0006, -0.0005, +0.0000,
-	-0.0011, +0.0002, +0.0007, -0.0004, +0.0001, +0.0002, -0.0014, +0.0002, +0.0003,
-	//45
-	-0.0374, -0.0364, +0.0626, -0.1302, -0.0309, -0.1094, +0.2483, -0.0440, +0.0841,
-	+0.0437, +0.0563, +0.0163, -0.0785, +0.3315, -0.1671, +0.0044, -0.3191, +0.1414,
-	+0.0616, -0.0526, +0.1331, -0.0226, +0.1955, -0.2987, +0.2956, +0.0508, -0.2351,
-	-0.1716, -0.0640, -0.0233, -0.3280, +0.0467, -0.3148, +0.2626, +0.4428, -0.1073,
-	-0.0010, +0.1931, +0.2608, -0.2589, +0.1616, -0.1773, -0.0218, +0.3485, +0.0865,
-	-0.0457, -0.0391, +0.1322, -0.2080, +0.0389, -0.2736, -0.1092, +0.0300, -0.1196,
-	+0.0442, -0.0213, -0.0125, -0.1207, +0.2580, -0.0946, -0.0196, -0.1229, +0.0918,
-	+0.0777, -0.0750, +0.0416, -0.1104, +0.1462, +0.0046, +0.2114, -0.0638, +0.0129,
-	-0.0203, -0.0560, +0.1633, -0.1222, +0.0076, -0.0904, -0.0162, +0.0453, -0.0171,
-	//46
-	-0.0060, -0.0525, +0.0377, +0.0081, +0.0289, -0.2535, -0.1002, +0.0401, -0.0080,
-	+0.0395, +0.0648, +0.1046, -0.0570, +0.1222, +0.3196, +0.2824, -0.1657, +0.1464,
-	-0.0116, -0.1224, -0.0105, +0.0802, -0.1704, -0.2288, +0.0538, -0.0263, -0.1205,
-	+0.0063, +0.1002, -0.0848, +0.0278, -0.0626, -0.2853, -0.0477, +0.3000, -0.0348,
-	-0.1589, +0.2223, -0.0811, -0.1661, +0.2328, +0.1975, -0.0764, +0.0109, -0.1202,
-	-0.1356, +0.1282, +0.0168, -0.3164, -0.0303, +0.1239, -0.0259, -0.0192, -0.0061,
-	-0.0318, -0.1741, +0.1495, -0.0962, -0.1037, +0.0249, +0.0517, -0.2208, +0.0726,
-	+0.2380, -0.1129, +0.2162, +0.1971, +0.1128, +0.0596, +0.2132, -0.1856, +0.2207,
-	-0.1044, +0.1277, -0.1444, -0.0924, +0.0553, -0.1869, +0.0999, -0.0670, -0.0445,
-	//47
-	+0.0006, +0.0013, -0.0012, -0.0003, +0.0006, +0.0002, -0.0010, +0.0002, -0.0011,
-	+0.0003, +0.0002, -0.0013, +0.0002, -0.0022, -0.0005, -0.0004, +0.0009, +0.0006,
-	-0.0006, +0.0014, -0.0005, +0.0011, +0.0001, +0.0011, -0.0001, -0.0002, -0.0000,
-	-0.0016, +0.0002, -0.0012, -0.0002, +0.0010, -0.0006, -0.0002, +0.0003, -0.0001,
-	+0.0002, +0.0015, -0.0000, -0.0004, +0.0007, -0.0003, -0.0009, -0.0001, -0.0015,
-	+0.0000, -0.0015, -0.0005, +0.0003, +0.0004, +0.0009, -0.0014, +0.0002, -0.0009,
-	+0.0005, +0.0003, +0.0003, -0.0003, +0.0007, +0.0014, -0.0003, -0.0008, +0.0011,
-	-0.0008, -0.0004, -0.0007, -0.0014, -0.0006, -0.0002, +0.0015, +0.0006, -0.0001,
-	+0.0006, +0.0004, -0.0009, -0.0007, -0.0003, +0.0012, +0.0011, -0.0016, -0.0023,
-	//48
-	+0.1728, -0.1200, -0.1469, +0.3504, +0.1332, +0.0061, +0.0947, +0.4392, -0.2194,
-	-0.1411, -0.0742, -0.0202, -0.2831, -0.7260, -0.4555, -0.3006, -0.2197, -0.0439,
-	+0.1006, -0.0583, +0.0940, +0.2225, +0.0854, +0.1545, -0.0827, -0.1191, +0.1371,
-	+0.0057, +0.0993, -0.0013, +0.1365, +0.4763, +0.7071, +0.1989, -0.0854, +0.2161,
-	+0.0507, -0.1780, -0.1850, +0.1194, -0.1992, -0.1366, -0.0785, -0.0530, -0.0517,
-	+0.2210, -0.1188, -0.0610, +0.3060, -0.2749, -0.2652, +0.0759, +0.2721, -0.1150,
-	-0.1472, -0.0284, -0.0587, +0.0676, -0.1320, +0.0678, -0.0195, -0.1230, -0.0900,
-	-0.0484, +0.2151, +0.0169, +0.0333, +0.1068, +0.2423, +0.0506, -0.0797, +0.1420,
-	+0.0201, +0.0279, -0.1290, -0.0095, -0.0722, -0.1198, +0.0971, -0.1120, +0.0449,
-	//49
-	+0.0848, -0.0782, -0.1043, +0.1760, -0.1539, -0.0407, -0.0448, +0.3480, -0.2108,
-	-0.1250, +0.0969, -0.1383, +0.1082, +0.1017, -0.0166, -0.3548, -0.0069, +0.0964,
-	+0.0819, +0.0499, +0.1003, -0.0598, +0.0293, +0.4506, +0.1569, -0.2594, +0.1811,
-	+0.0635, -0.1638, +0.0857, -0.2160, -0.4133, -0.0723, +0.2215, -0.2205, +0.0278,
-	-0.0838, -0.0219, +0.3527, -0.0203, -0.2202, -0.2129, +0.1432, +0.0907, -0.1254,
-	+0.0234, -0.2012, +0.1566, +0.2480, +0.1099, -0.3826, -0.2578, +0.2733, -0.0955,
-	+0.3325, -0.2187, -0.4021, +0.2550, +0.5937, +0.2013, -0.0608, +0.0296, +0.1294,
-	-0.2539, +0.4650, -0.0404, -0.4773, -0.2848, +0.2024, +0.1172, -0.2867, -0.0128,
-	-0.1086, +0.0640, +0.0608, +0.0259, -0.0488, -0.0095, +0.0364, -0.0102, +0.0635,
-	//50
-	+0.0503, +0.0373, +0.0270, -0.1651, +0.0585, +0.1514, -0.1213, -0.1602, -0.0122,
-	-0.0927, -0.1672, -0.0887, +0.0381, +0.0407, -0.0318, -0.0435, +0.2124, +0.1566,
-	+0.1170, +0.1298, +0.1561, +0.2374, +0.0756, -0.0214, -0.0955, +0.0423, -0.0536,
-	-0.0133, +0.0185, -0.1462, -0.0794, -0.1286, -0.1157, -0.1403, -0.0265, -0.0014,
-	-0.0526, +0.0287, -0.1631, -0.0229, +0.0966, +0.0925, -0.0390, -0.0071, -0.0109,
-	+0.1347, +0.0054, -0.0688, +0.1059, +0.1839, +0.0960, +0.0323, +0.1027, +0.0381,
-	+0.1305, -0.1964, -0.0737, +0.0238, -0.1524, -0.3187, -0.1089, +0.0504, -0.0743,
-	+0.0186, -0.1691, +0.2006, +0.1918, -0.1348, -0.1520, +0.2169, +0.1877, -0.0370,
-	+0.0243, -0.0322, -0.0121, -0.0059, +0.0094, +0.0648, +0.0223, -0.2178, +0.0209,
-	//51
-	-0.0000, +0.0006, +0.0010, -0.0014, -0.0011, -0.0015, +0.0001, +0.0000, -0.0006,
-	-0.0002, +0.0002, -0.0006, -0.0005, +0.0008, -0.0016, -0.0001, +0.0014, -0.0007,
-	-0.0003, -0.0003, +0.0011, +0.0017, -0.0027, +0.0005, +0.0017, -0.0013, +0.0004,
-	+0.0012, +0.0005, -0.0014, +0.0001, +0.0001, -0.0018, -0.0004, -0.0012, +0.0001,
-	+0.0005, -0.0015, -0.0004, -0.0008, -0.0014, -0.0014, -0.0009, -0.0020, +0.0001,
-	-0.0002, -0.0001, +0.0003, +0.0006, -0.0016, +0.0003, +0.0002, +0.0009, +0.0010,
-	-0.0008, -0.0004, -0.0016, +0.0003, +0.0012, -0.0001, -0.0001, +0.0007, -0.0012,
-	+0.0026, +0.0003, -0.0020, +0.0011, +0.0009, +0.0007, -0.0012, -0.0014, +0.0005,
-	-0.0009, +0.0012, -0.0019, -0.0008, -0.0010, -0.0008, -0.0002, -0.0010, +0.0019,
-	//52
-	+0.0668, -0.1016, -0.2302, +0.0477, -0.1149, -0.0478, +0.0118, +0.0855, -0.1606,
-	+0.1001, +0.2036, +0.0983, +0.2252, +0.0402, -0.2340, -0.0214, +0.0931, +0.0813,
-	-0.1368, +0.0791, +0.0973, +0.1506, +0.1623, +0.0729, +0.0766, -0.2216, +0.1725,
-	-0.2092, -0.1829, -0.2614, -0.0510, +0.0863, -0.1124, +0.0991, +0.0680, -0.0485,
-	+0.0784, -0.0695, -0.0502, +0.0089, +0.1600, -0.0490, +0.0451, +0.0325, -0.1156,
-	+0.2617, -0.2127, +0.0143, +0.0285, -0.0338, -0.3003, -0.1557, +0.1459, -0.0107,
-	+0.1470, -0.2126, +0.0889, +0.0506, +0.0782, -0.0491, -0.0577, +0.1101, -0.1152,
-	-0.1017, +0.0060, +0.1484, -0.2315, +0.0216, +0.3726, +0.1251, -0.1767, +0.1372,
-	-0.0157, +0.0930, -0.0471, -0.0198, -0.0026, -0.1766, -0.0064, +0.0174, -0.0010,
-	//53
-	-0.0941, -0.0474, -0.0062, -0.0855, -0.1001, -0.0121, -0.0032, -0.0131, +0.0910,
-	+0.1012, +0.0469, +0.0674, +0.0563, -0.0075, -0.0122, +0.0178, -0.0439, -0.0851,
-	+0.0769, -0.0366, -0.0260, +0.0200, -0.0585, -0.0898, +0.0270, -0.0052, -0.1005,
-	+0.0290, -0.0433, -0.0402, +0.0323, +0.0403, +0.0596, +0.1583, +0.1066, +0.0834,
-	-0.0425, -0.0508, -0.0858, -0.0341, +0.0702, +0.1330, +0.1129, -0.0514, +0.0052,
-	-0.0579, -0.0125, -0.0383, -0.0389, +0.0315, +0.0896, +0.0190, -0.1440, -0.0293,
-	-0.0155, +0.0036, +0.0189, +0.0156, +0.0093, +0.0001, -0.0377, -0.0762, +0.0367,
-	+0.0138, -0.0276, +0.0245, +0.0609, +0.0417, -0.0200, -0.0253, -0.0018, -0.0181,
-	+0.0617, -0.0487, -0.0416, -0.0557, -0.0328, +0.0057, +0.0395, +0.0464, -0.0539,
-	//54
-	+0.0492, +0.1641, -0.1324, -0.0894, +0.2117, -0.0594, -0.1503, +0.0750, +0.0324,
-	-0.0338, -0.1085, -0.0698, -0.0308, -0.0027, -0.0624, +0.0394, -0.0512, -0.1203,
-	+0.0291, -0.0209, -0.1501, -0.1728, +0.0547, -0.0722, +0.0499, +0.0625, -0.1063,
-	-0.0283, +0.0215, -0.0562, -0.1394, +0.3558, +0.0117, +0.1128, +0.1567, -0.1720,
-	+0.1520, +0.0595, +0.1364, +0.3879, +0.9080, +0.3000, +0.2609, +0.0983, +0.0263,
-	-0.0607, +0.0716, +0.0335, +0.0498, +0.3100, -0.2205, -0.1712, -0.1091, +0.0192,
-	-0.1577, +0.1379, +0.0145, -0.0697, +0.0917, -0.0718, -0.0424, -0.0026, +0.0928,
-	-0.0812, +0.0676, -0.0831, -0.1606, +0.0079, +0.0782, +0.0879, -0.0797, -0.0954,
-	+0.2028, -0.1759, -0.0532, +0.0528, +0.1626, -0.1341, -0.0750, +0.1195, -0.0294,
-	//55
-	-0.0434, +0.1374, +0.1169, -0.1157, +0.0291, -0.0031, -0.0029, -0.1910, +0.0506,
-	-0.1251, -0.0331, +0.0028, -0.3155, +0.0153, +0.1685, +0.0884, -0.1169, +0.1330,
-	+0.1565, -0.0103, +0.1513, +0.1989, +0.2397, +0.0392, +0.1015, +0.0149, +0.0109,
-	-0.1196, -0.0759, -0.0819, -0.0998, -0.2372, -0.5238, -0.1418, -0.0526, -0.1128,
-	-0.0378, +0.2154, +0.0826, +0.1113, +0.3109, +0.0528, +0.1570, +0.0942, +0.0648,
-	-0.1116, +0.0143, -0.1510, -0.4222, -0.2300, -0.0728, +0.0733, -0.1276, -0.0405,
-	+0.1079, -0.2848, -0.0871, +0.0566, +0.0899, -0.0136, +0.1002, -0.1488, -0.0209,
-	+0.1635, -0.1146, +0.3254, +0.4442, +0.4076, -0.1654, +0.1203, +0.1435, +0.1671,
-	+0.0281, -0.0659, +0.0727, -0.3280, +0.0715, -0.1375, +0.1820, -0.1303, -0.0714,
-	//56
-	-0.0749, +0.2190, +0.1223, -0.1253, +0.2718, +0.1557, -0.4426, -0.0216, +0.0816,
-	-0.0605, -0.0358, -0.1073, -0.0262, -0.2389, +0.0046, +0.2880, +0.0515, -0.0807,
-	+0.0732, -0.0865, -0.0842, +0.0810, -0.5388, -0.0456, +0.2504, +0.1195, -0.1465,
-	+0.1283, -0.0190, -0.1570, +0.6802, +0.2139, +0.1021, -0.3662, +0.3152, -0.1335,
-	-0.0433, +0.0696, -0.7429, +0.1356, +0.0795, +0.1766, -0.6118, -0.0358, +0.1516,
-	-0.0336, +0.4145, -0.0440, +0.0862, +0.0193, +0.6814, +0.0529, -0.1763, +0.0520,
-	-0.0997, -0.0272, -0.0113, -0.0501, -0.5713, -0.1305, -0.0099, +0.0143, +0.0951,
-	-0.0552, -0.0040, +0.1081, +0.3577, -0.0376, -0.0700, -0.0230, -0.0970, -0.0267,
-	+0.0629, +0.0654, -0.2827, -0.0413, +0.1565, +0.0039, +0.0103, +0.1934, -0.0889,
-	//57
-	+0.0892, -0.0891, -0.0189, +0.0682, +0.0186, -0.0816, -0.0211, +0.0000, -0.0843,
-	+0.0866, -0.0156, +0.1282, +0.1115, +0.0228, +0.0288, +0.0987, +0.0221, +0.0006,
-	+0.0052, -0.0618, +0.0386, +0.0135, +0.0263, -0.0229, -0.0046, -0.0202, -0.0024,
-	-0.0065, -0.0746, +0.0111, -0.0161, +0.0585, -0.0408, +0.0616, +0.0761, +0.0885,
-	+0.0426, -0.0377, +0.0812, +0.1881, +0.2068, +0.0272, +0.1705, +0.0741, +0.0690,
-	+0.0169, -0.0681, -0.0206, +0.0926, +0.0851, -0.1353, -0.0092, -0.0386, -0.0374,
-	-0.0131, +0.0088, -0.0167, +0.0323, +0.0924, +0.0033, +0.0402, +0.0064, +0.0492,
-	-0.0573, +0.0347, -0.0110, -0.0768, +0.0365, +0.0360, +0.0084, -0.0396, +0.0825,
-	-0.0684, -0.1115, +0.0353, -0.0065, +0.0662, -0.0018, +0.0177, +0.0060, +0.0370,
-	//58
-	-0.0025, +0.1702, +0.0476, +0.2458, +0.0568, -0.1805, -0.2669, +0.2433, -0.1265,
-	-0.0761, -0.0885, -0.2999, -0.1094, -0.0367, +0.1299, -0.0103, +0.0557, +0.0030,
-	+0.1120, +0.0404, -0.1048, +0.2384, +0.0718, +0.1048, +0.1929, +0.0405, -0.0746,
-	-0.1014, +0.0437, -0.1188, +0.1275, -0.1313, -0.1005, -0.0931, -0.1134, +0.0443,
-	+0.1209, -0.0498, +0.0236, +0.2035, -0.1811, +0.0123, -0.0384, +0.0303, +0.1583,
-	+0.0208, -0.1011, -0.1542, -0.0936, -0.2476, +0.2511, +0.0580, -0.1719, -0.0630,
-	-0.0802, +0.1810, +0.0487, +0.0144, +0.0553, +0.3152, -0.1498, -0.0748, +0.2076,
-	-0.0744, +0.0764, +0.0456, +0.2617, -0.1135, -0.1585, -0.3202, -0.0583, -0.1838,
-	-0.0167, +0.1650, -0.3521, +0.0594, -0.1485, +0.3126, +0.2443, +0.0457, +0.0480,
-	//59
-	+0.0345, -0.1739, -0.0262, +0.0541, +0.0811, -0.0499, -0.0069, +0.0450, -0.0317,
-	+0.0814, +0.0138, +0.1726, +0.0647, +0.0534, +0.1124, +0.1764, +0.0980, +0.1365,
-	-0.0783, -0.0956, +0.0186, -0.0775, +0.0169, -0.0049, -0.0122, -0.0549, -0.0010,
-	-0.0081, -0.0649, +0.0412, -0.0456, +0.1271, +0.0359, +0.0716, -0.0042, +0.0185,
-	+0.0963, -0.0650, +0.0571, +0.1384, +0.2540, +0.0570, +0.1576, +0.0252, +0.0414,
-	-0.0327, -0.1244, -0.0699, +0.0065, +0.0908, -0.1443, -0.0310, -0.0660, -0.1010,
-	+0.0390, +0.0746, -0.0118, -0.0328, +0.0755, -0.0147, +0.0511, +0.0267, -0.0094,
-	+0.0780, +0.1630, +0.0361, -0.1005, -0.0134, -0.0432, +0.0469, +0.0623, +0.0746,
-	-0.0980, -0.1431, +0.0047, -0.0300, +0.0516, -0.0903, +0.0043, +0.0453, +0.0288,
-	//60
-	-0.1336, -0.0374, -0.0053, -0.0183, +0.0383, +0.0579, +0.0263, -0.0224, -0.0753,
-	-0.0382, +0.0504, +0.0348, +0.0303, +0.0781, +0.0266, -0.0554, -0.0456, -0.0043,
-	+0.0012, +0.0810, +0.0072, -0.0099, +0.0855, +0.0370, -0.1047, -0.0906, -0.0594,
-	+0.0193, +0.0723, -0.0087, -0.0261, +0.0988, +0.0744, -0.0628, -0.0631, -0.0453,
-	+0.0471, +0.0118, -0.0120, +0.0717, +0.1257, +0.0720, +0.0277, +0.0361, +0.0177,
-	+0.0315, -0.0479, -0.0600, +0.0384, +0.0460, -0.0015, +0.0180, +0.0653, +0.0107,
-	+0.0634, -0.0198, -0.1005, -0.0360, +0.0045, +0.0171, -0.0010, +0.0194, +0.0258,
-	+0.0323, -0.0068, -0.0905, -0.0237, +0.0717, +0.0874, +0.0139, +0.0061, +0.0618,
-	+0.0063, +0.0274, +0.0061, -0.0159, +0.0101, -0.0058, -0.0267, +0.0026, +0.0491,
-	//61
-	+0.0625, -0.0109, -0.0347, +0.2528, -0.2762, +0.2485, -0.4919, +0.4779, -0.1779,
-	-0.0844, +0.1312, -0.4011, +0.1888, -0.0940, +0.4499, -0.3327, +0.0334, +0.0306,
-	-0.0596, +0.2729, -0.1741, +0.2182, -0.3480, +0.2393, -0.1159, -0.2200, +0.1572,
-	-0.2118, +0.2853, -0.0022, -0.0775, -0.1874, +0.7579, +0.0819, -0.6559, +0.2335,
-	-0.0375, +0.0362, -0.0233, -0.1263, -0.3289, +0.4651, +0.1472, -0.4950, +0.0171,
-	+0.1320, -0.1742, -0.0143, +0.2039, -0.2496, -0.0964, +0.0904, +0.2652, -0.1964,
-	+0.0223, +0.1061, -0.0432, +0.1733, +0.0231, +0.0304, -0.1914, +0.3014, +0.0145,
-	-0.2056, +0.1115, -0.1016, -0.0557, -0.0033, +0.1384, -0.1619, -0.0043, -0.1720,
-	+0.0983, -0.0531, -0.0020, +0.0854, -0.0973, +0.0345, +0.0755, -0.0114, +0.0757,
-	//62
-	+0.0290, +0.1637, -0.0745, -0.0675, +0.0668, +0.0297, -0.1010, -0.0242, +0.1266,
-	-0.0688, -0.1298, -0.2025, -0.1996, -0.1180, -0.0749, -0.0428, +0.0036, -0.0762,
-	+0.0502, +0.0027, +0.0271, +0.0474, +0.0898, +0.1038, +0.1211, +0.0998, -0.0582,
-	+0.0548, +0.0341, -0.0022, -0.0120, +0.0279, +0.0015, +0.0154, +0.0417, -0.0921,
-	-0.0477, +0.0686, +0.0511, -0.0210, +0.0752, +0.0525, -0.0261, -0.0471, -0.0502,
-	-0.1368, +0.0633, +0.1550, +0.1697, +0.2602, +0.1952, +0.0998, +0.0124, +0.0848,
-	-0.1625, +0.0487, +0.1465, +0.1433, +0.1616, +0.0512, +0.0654, +0.0249, +0.0116,
-	-0.0562, +0.0774, +0.0759, +0.0260, +0.0056, -0.0795, -0.0586, -0.0398, -0.0664,
-	+0.0979, -0.0031, -0.1057, -0.0003, +0.0182, -0.0560, -0.0462, +0.0877, +0.0760,
-	//63
-	+0.0347, -0.1416, -0.1128, +0.1999, -0.0949, -0.2378, -0.0675, +0.2511, -0.1187,
-	+0.1192, +0.1516, -0.0195, +0.0298, +0.1830, +0.2252, +0.0594, -0.0415, +0.0680,
-	-0.0991, -0.1341, -0.1139, +0.1177, +0.0211, -0.2812, +0.0747, +0.0687, -0.1960,
-	+0.0718, +0.1183, -0.0007, -0.0948, -0.1025, -0.2501, +0.1760, +0.0936, +0.0904,
-	-0.1120, +0.0874, -0.1395, -0.3595, +0.3106, +0.0456, -0.2439, -0.1720, -0.0778,
-	+0.0732, +0.0151, +0.1762, +0.0828, +0.4929, +0.2105, -0.1395, +0.3501, +0.0636,
-	+0.1103, +0.0283, +0.2762, +0.1404, -0.0632, +0.1369, -0.1415, +0.1467, -0.0446,
-	-0.4688, +0.2780, -0.2456, -0.5055, -0.5325, -0.0105, -0.2260, -0.1170, -0.1383,
-	+0.0941, +0.0421, -0.1627, +0.3231, +0.0169, +0.1862, +0.0404, -0.0873, +0.1451,
-	//64
-	+0.0028, -0.0265, +0.0122, +0.0436, -0.0678, -0.0886, -0.0434, -0.0408, +0.0292,
-	+0.0132, -0.0514, +0.0153, +0.0689, +0.0022, -0.0308, -0.0025, +0.0136, +0.0435,
-	+0.0330, -0.0197, +0.0349, +0.0988, +0.1310, +0.0874, +0.0130, -0.0096, -0.0398,
-	-0.0091, -0.0436, -0.0133, +0.0313, +0.0561, +0.0137, +0.0123, +0.0175, +0.0143,
-	-0.0378, -0.0337, +0.0379, +0.1242, +0.0562, +0.0116, +0.1186, +0.0745, +0.0008,
-	+0.0591, -0.0402, -0.0038, +0.1507, +0.0813, -0.0060, +0.0924, +0.0692, -0.0138,
-	+0.0367, -0.0308, -0.0444, +0.0798, +0.1092, +0.0564, +0.0347, +0.0011, +0.0264,
-	-0.0329, -0.0225, -0.0761, -0.0277, +0.1030, +0.1069, -0.0249, -0.0697, +0.0882,
-	+0.0077, +0.0186, +0.0150, -0.0538, -0.0238, +0.0411, +0.0359, +0.0081, +0.0821,
-};
-
-#endif //__CONVOLUTION1_H_INCLUDED
