--- _Old_Version/convolution2.h
@@ -1,343 +1,0 @@
-#ifndef __CONVOLUTION2_H_INCLUDED
-#define __CONVOLUTION2_H_INCLUDED
-
-#include <cv.h>
-#include <highgui.h>
-
-#define CONV1_FILTERS	64
-#define CONV2_FILTERS	32
-
-void Convolution11(IplImage* src[], IplImage* dst, float kernel[CONV1_FILTERS], float bias)
-{
-	/*遍历dst所有像素*/
-	for (int row = 0; row < dst->height; row++)
-	{
-		for (int col = 0; col < dst->width; col++)
-		{
-			//对dst每个像素操作
-			double temp = 0;	//存储dst[row][col]像素值16
-			for (int i = 0; i < CONV1_FILTERS; i++)
-			{
-				unsigned int temppixel = 0;	//取出src[i].[row][col]像素值
-				temppixel += (unsigned char)(src[i]->imageData + row*src[i]->widthStep)[2 * col];
-				temppixel <<= 8;
-				temppixel += (unsigned char)(src[i]->imageData + row*src[i]->widthStep)[2 * col + 1];
-
-				temp += temppixel * kernel[i];
-			}
-			temp += bias;
-
-			unsigned int tempINT = 0;
-			if (temp > 0)
-			{
-				tempINT = (unsigned int)(temp + 0.5);
-			}
-
-			(dst->imageData + row*dst->widthStep)[2 * col + 1] = (unsigned char)tempINT;
-			(dst->imageData + row*dst->widthStep)[2 * col] = (unsigned char)(tempINT >> 8);
-			//对dst每个像素操作
-		}
-	}
-}
-
-//偏置32
-float biases_conv2[CONV2_FILTERS] = {
-	31.4415, 33.8385, 18.0540, 13.1325, 14.8920, 33.5325, 20.6550, 25.8825,
-	14.2545, 13.3365, -28.5855, 19.0230, -26.2650, 27.5910, 40.0095, 46.4865,
-	17.2890, -7.1145, 27.9480, 8.7465, 15.2745, -12.2910, 19.0740, -18.7170,
-	-5.6355, -12.6225, 0.0510, 42.8145, -22.0320, 15.4275, -44.2170, 67.7025,
-};
-
-//权重32*64
-float weights_conv2_data[CONV2_FILTERS][CONV1_FILTERS] = {
-	//01
-	-0.2400, +0.1037, +0.1117, +0.0435, -0.2373, +0.1215, -0.2368, -0.2855,
-	-0.6127, -0.1222, -0.0601, +0.1851, +0.1601, +0.1392, +0.2136, +0.1079,
-	-0.2700, -0.1424, -0.0014, -0.0050, +0.3402, +0.0006, +0.0000, +0.0038,
-	+0.0012, +0.0017, -0.0012, +0.0499, -0.0001, -0.1774, +0.0011, -0.2238,
-	+0.0036, -0.3211, -0.1617, -0.0008, +0.0506, +0.2107, +0.0772, -0.3316,
-	+0.0627, +0.0003, -0.0013, +0.0004, -0.4159, -0.0200, +0.0003, +0.1791,
-	-0.5951, +0.3260, +0.0014, -0.0893, -0.1307, +0.2111, -0.0585, -0.1194,
-	+0.0856, -0.1141, +0.2838, -0.0116, +0.3210, -0.2009, -0.3413, -0.0292,
-	//02
-	-0.1529, -0.4510, -0.0708, +0.1441, +0.0595, -0.0366, +0.3870, -0.1276,
-	+0.0600, +0.0054, +0.1643, -0.0425, +0.0330, -0.1997, -0.2559, -0.0548,
-	-0.1335, +0.0519, -0.0006, +0.1412, -0.0660, +0.0003, +0.0005, +0.0019,
-	-0.0004, +0.0006, -0.0008, +0.0088, +0.0003, -0.0316, -0.0011, +0.5397,
-	+0.1684, +0.1002, -0.0446, -0.0008, +0.1456, +0.0537, +0.3999, +0.3651,
-	-0.0001, -0.0016, +0.0002, +0.0015, +0.1720, +0.2043, +0.0007, -0.1559,
-	+0.1317, +0.0430, +0.0005, -0.1449, -0.0563, +0.0406, +0.1233, -0.0687,
-	+0.1168, +0.2059, -0.0764, -0.1059, -0.2586, -0.2342, +0.1204, -0.0915,
-	//03
-	+0.1729, +0.1209, +0.1217, -0.1777, -0.0384, +0.1379, -0.0180, -0.0035,
-	+0.0524, -0.0137, -0.0238, +0.0875, +0.0095, +0.0022, +0.0055, +0.1434,
-	-0.3052, -0.0277, +0.0003, -0.0909, +0.0800, +0.0003, +0.0001, +0.0033,
-	-0.0010, +0.0010, +0.0024, +0.1591, +0.0007, -0.2398, +0.0004, -0.1528,
-	+0.0969, -0.1340, -0.1083, -0.0011, -0.0791, -0.0455, +0.0714, -0.1195,
-	+0.0652, +0.0005, -0.0012, +0.0002, -0.0219, -0.1097, -0.0004, -0.0620,
-	-0.0219, +0.1316, +0.0011, -0.1214, -0.0627, +0.2554, +0.1002, +0.1312,
-	+0.1652, -0.1493, +0.1441, +0.0394, +0.1944, -0.0161, -0.0876, +0.0857,
-	//04
-	-0.0347, +0.0888, +0.1153, -0.2509, -0.1818, +0.1117, -0.0645, -0.0473,
-	-0.1316, +0.0190, -0.2969, +0.3151, +0.0167, +0.0856, +0.0170, +0.1361,
-	-0.0086, -0.1584, -0.0006, +0.1460, +0.0294, -0.0021, -0.0022, +0.0011,
-	+0.0003, -0.0003, -0.0014, +0.4037, +0.0010, -0.0892, +0.0017, +0.0494,
-	+0.0305, -0.1099, +0.0418, -0.0004, -0.0917, -0.1121, +0.0182, -0.1032,
-	+0.0366, -0.0009, +0.0005, +0.0020, -0.1728, -0.1575, +0.0016, +0.1030,
-	-0.0563, +0.1217, -0.0003, -0.0463, -0.0520, +0.1792, +0.1714, +0.2015,
-	+0.1466, -0.0650, +0.0864, +0.0266, -0.2657, -0.1008, -0.2020, +0.0729,
-	//05
-	+0.2060, +0.1484, +0.1169, -0.2924, +0.0209, +0.1405, -0.0328, +0.0140,
-	+0.0489, -0.0856, +0.0728, +0.0699, -0.0065, +0.0847, +0.0138, +0.1044,
-	-0.3232, -0.0335, -0.0006, -0.2523, +0.0813, -0.0013, -0.0003, +0.0022,
-	+0.0007, +0.0009, -0.0005, +0.1803, +0.0005, -0.3850, -0.0007, -0.1552,
-	+0.0770, -0.1991, -0.0406, +0.0002, -0.0895, -0.1745, +0.1934, -0.0399,
-	-0.0535, +0.0010, +0.0013, -0.0000, +0.0465, -0.1538, -0.0009, +0.0683,
-	-0.0747, +0.1969, -0.0002, -0.0719, -0.0801, +0.2258, +0.1159, +0.1791,
-	+0.1320, -0.1111, +0.1246, +0.0064, +0.0932, +0.0328, +0.0406, +0.0763,
-	//06
-	+0.2665, +0.0070, +0.0569, -0.2168, +0.0778, +0.0551, -0.0776, +0.0387,
-	-0.7930, -0.1106, +0.1175, +0.1326, -0.0642, +0.0215, +0.1065, +0.0522,
-	-0.4888, -0.2284, +0.0011, +0.1531, +0.0793, +0.0004, +0.0018, +0.0007,
-	+0.0002, -0.0015, +0.0003, +0.1798, -0.0003, -0.0918, -0.0003, -0.1768,
-	+0.0535, -0.3016, +0.0194, -0.0001, -0.0371, -0.0780, -0.1688, -0.5796,
-	+0.1565, -0.0013, -0.0011, +0.0012, +0.1066, -0.1248, +0.0004, +0.0930,
-	-0.3883, +0.2995, +0.0008, -0.2824, -0.0977, +0.3056, +0.2115, +0.0701,
-	+0.0824, -0.0095, +0.1499, +0.0257, +0.2209, +0.0165, -0.0407, +0.0129,
-	//07
-	+0.0062, -0.0133, +0.1917, -0.0171, -0.0751, +0.1805, -0.3079, +0.0483,
-	+0.0596, +0.1261, -0.2460, +0.1237, +0.0286, -0.0062, -0.0489, +0.1681,
-	-0.1049, -0.1049, -0.0005, +0.2358, +0.0508, -0.0012, -0.0001, +0.0018,
-	+0.0009, -0.0015, -0.0005, +0.1632, +0.0020, +0.0158, +0.0016, -0.0639,
-	+0.0541, -0.1496, -0.0143, +0.0013, -0.1235, +0.0164, -0.2519, -0.0428,
-	+0.1909, -0.0002, +0.0019, -0.0005, -0.1110, -0.0506, -0.0011, -0.0023,
-	-0.1191, -0.0224, +0.0001, -0.1372, -0.0242, +0.2427, +0.1388, +0.1363,
-	+0.1573, -0.1580, +0.1089, +0.0772, +0.0974, -0.0917, -0.2062, +0.1189,
-	//08
-	-0.0951, -0.0754, +0.2706, +0.0879, +0.1014, +0.2685, -0.5909, +0.1982,
-	+0.0894, +0.0557, -0.1161, +0.0391, +0.0632, +0.1078, -0.0198, +0.1451,
-	-0.0609, -0.0701, +0.0002, +0.1747, +0.0606, -0.0001, +0.0006, +0.0018,
-	+0.0011, +0.0002, +0.0020, +0.0763, +0.0003, -0.0005, -0.0001, -0.0320,
-	+0.0418, -0.3185, +0.0827, +0.0022, -0.1668, -0.0318, -0.3289, +0.0465,
-	+0.1761, +0.0007, +0.0001, -0.0002, -0.0969, -0.0332, -0.0008, +0.2259,
-	-0.2223, -0.0546, -0.0001, -0.1492, -0.0390, +0.2070, +0.0770, +0.1661,
-	+0.1441, -0.2381, +0.1024, +0.0977, +0.0884, -0.0241, +0.0007, +0.1787,
-	//09
-	+0.1167, +0.1086, +0.1134, -0.2323, -0.0541, +0.1293, -0.0307, -0.0236,
-	-0.0240, -0.0223, -0.1005, +0.1605, -0.0048, +0.0409, +0.0047, +0.1292,
-	-0.2130, -0.0727, +0.0003, -0.0366, +0.0591, +0.0006, -0.0004, +0.0011,
-	+0.0010, -0.0012, -0.0003, +0.2431, +0.0009, -0.2085, +0.0004, -0.0737,
-	+0.0682, -0.1429, -0.0299, -0.0008, -0.0934, -0.1140, +0.0705, -0.1140,
-	+0.0276, -0.0014, -0.0007, +0.0007, -0.0532, -0.1442, -0.0019, +0.0361,
-	-0.0455, +0.1394, +0.0002, -0.0952, -0.0658, +0.2246, +0.1280, +0.1765,
-	+0.1518, -0.1207, +0.1208, +0.0237, +0.0076, -0.0405, -0.1098, +0.0782,
-	//10
-	+0.0611, +0.1183, +0.1044, -0.2634, -0.0553, +0.1266, -0.0722, -0.0095,
-	-0.0934, -0.0803, -0.0920, +0.1974, +0.0019, +0.1129, +0.0603, +0.1073,
-	-0.1505, -0.0766, +0.0009, -0.0465, +0.0484, +0.0003, +0.0001, +0.0008,
-	+0.0003, +0.0023, -0.0005, +0.2892, +0.0005, -0.2450, -0.0007, -0.0164,
-	+0.0581, -0.1963, +0.0228, -0.0005, -0.0906, -0.1478, +0.1193, -0.0964,
-	-0.0123, +0.0003, -0.0010, +0.0007, -0.1043, -0.1757, -0.0005, +0.1477,
-	-0.0458, +0.1693, -0.0022, -0.0735, -0.0814, +0.1803, +0.1127, +0.2046,
-	+0.1401, -0.1239, +0.1052, +0.0108, -0.1170, -0.0355, -0.0512, +0.0757,
-	//11
-	+0.1935, +0.2059, +0.0790, -0.3367, -0.3291, +0.1312, -0.1209, -0.1102,
-	-0.0459, -0.0584, -0.1608, +0.1280, -0.0689, +0.2197, +0.1431, +0.0569,
-	-0.1542, -0.0902, -0.0012, +0.0959, +0.0433, -0.0001, +0.0018, +0.0007,
-	-0.0003, -0.0005, +0.0007, +0.2549, +0.0005, -0.1634, +0.0002, -0.1463,
-	-0.0529, -0.0264, -0.3561, +0.0010, -0.0727, +0.0890, +0.3606, -0.0372,
-	+0.1762, -0.0004, +0.0016, -0.0004, -0.0713, -0.1050, +0.0011, -0.0016,
-	-0.3988, +0.1345, +0.0006, -0.0331, -0.0451, +0.1176, +0.1592, +0.1835,
-	+0.0170, -0.1262, +0.0323, -0.0703, +0.1437, +0.0241, -0.1478, +0.0310,
-	//12
-	+0.0181, +0.0032, +0.1705, -0.0665, -0.0119, +0.1775, -0.2487, +0.0448,
-	-0.0421, +0.0545, -0.2000, +0.1537, +0.0264, +0.0293, -0.0125, +0.1489,
-	-0.1099, -0.0809, +0.0001, +0.1519, +0.0471, +0.0002, +0.0006, +0.0015,
-	-0.0004, +0.0010, -0.0017, +0.2006, -0.0009, -0.0506, -0.0015, -0.0059,
-	+0.0586, -0.1791, +0.0560, -0.0010, -0.1198, -0.0480, -0.1832, -0.0923,
-	+0.1218, -0.0011, -0.0004, +0.0028, -0.1072, -0.0943, +0.0009, +0.0982,
-	-0.0938, +0.0393, +0.0005, -0.1312, -0.0496, +0.2264, +0.1142, +0.1931,
-	+0.1580, -0.1570, +0.1133, +0.0571, -0.0193, -0.0762, -0.1904, +0.1157,
-	//13
-	+0.2987, +0.2621, +0.0740, -0.2949, -0.2418, +0.1759, -0.2775, -0.1019,
-	-0.0899, -0.1453, -0.0378, +0.0549, -0.0882, +0.3060, +0.0213, +0.1116,
-	-0.2807, -0.0838, -0.0003, +0.2570, -0.0937, +0.0003, -0.0016, +0.0034,
-	-0.0012, +0.0002, +0.0016, +0.3054, +0.0015, -0.0811, +0.0004, -0.2963,
-	-0.0845, -0.2335, -0.3916, -0.0034, -0.2629, -0.1144, +0.3470, +0.0152,
-	+0.2791, -0.0001, -0.0015, -0.0015, -0.2259, -0.2607, +0.0002, +0.2502,
-	-0.2572, +0.3346, -0.0002, -0.1138, -0.1319, +0.0694, +0.2192, +0.2128,
-	+0.0628, +0.0102, -0.0485, -0.0643, +0.1578, -0.0852, -0.0553, +0.0765,
-	//14
-	-0.0001, +0.0467, -0.0278, +0.0150, +0.0420, +0.0312, +0.0107, +0.0278,
-	-0.1254, -0.0287, +0.0212, +0.0249, -0.1517, +0.0052, +0.1048, -0.0002,
-	+0.0061, +0.0503, -0.0007, +0.0578, -0.1095, +0.0002, -0.0001, -0.0013,
-	+0.0020, -0.0014, -0.0003, +0.0323, -0.0010, +0.0110, +0.0016, -0.0202,
-	-0.0731, -0.0317, -0.0644, +0.0003, -0.0430, +0.0712, +0.0397, +0.0003,
-	-0.0474, -0.0008, +0.0001, -0.0006, -0.0017, -0.0080, -0.0001, +0.1486,
-	-0.0214, +0.0069, +0.0008, +0.0317, -0.0502, +0.0722, -0.0906, +0.0737,
-	-0.0538, +0.0666, -0.2125, +0.0160, -0.0094, +0.1115, -0.0006, +0.0473,
-	//15
-	+0.7820, +0.3796, +0.0327, -0.2809, +0.2235, -0.1240, -0.4067, -0.3048,
-	-0.3514, -0.2315, -0.1011, +0.3373, -0.2702, +0.2750, -0.0988, +0.1319,
-	+0.0730, -0.1980, +0.0011, +0.1960, +0.1611, +0.0006, +0.0005, +0.0044,
-	+0.0007, -0.0002, -0.0016, +0.2107, +0.0006, -0.1406, -0.0017, -0.1032,
-	+0.1223, +0.0531, -0.2396, -0.0012, +0.0270, -0.1469, +0.4273, +0.0968,
-	-0.1674, -0.0005, +0.0000, +0.0015, -0.0990, -0.0252, +0.0009, +0.0568,
-	-0.5185, +0.1067, -0.0002, -0.0073, -0.1319, +0.2360, +0.2032, +0.5668,
-	-0.0751, -0.2078, -0.1193, +0.2009, +0.6165, -0.0496, -0.0225, -0.0289,
-	//16
-	+0.2119, +0.2405, -0.0770, -0.1388, -0.0555, -0.2376, +0.2857, +0.0399,
-	-0.9207, -0.0140, -0.1273, +0.3953, +0.0707, +0.1806, +0.1761, +0.0533,
-	-0.3085, -0.3814, +0.0010, +0.3004, +0.1906, -0.0006, +0.0019, +0.0033,
-	+0.0001, -0.0015, +0.0005, +0.2066, +0.0008, -0.1209, -0.0010, +0.0909,
-	+0.0569, -0.3447, -0.0947, +0.0002, -0.2950, -0.3668, +0.2186, -0.1417,
-	+0.4879, +0.0012, -0.0001, +0.0005, -0.1600, +0.0374, -0.0002, +0.3313,
-	-0.3065, +0.2708, +0.0001, -0.3078, -0.1309, +0.3704, +0.4220, +0.3843,
-	+0.0467, -0.0604, +0.2096, -0.0274, -0.0631, -0.1215, -0.5266, -0.2094,
-	//17
-	+0.0245, -0.0169, +0.1614, -0.0891, -0.0063, +0.1662, -0.2425, +0.0191,
-	-0.1148, +0.0958, -0.2490, +0.1546, +0.0088, +0.0340, -0.0445, +0.1308,
-	-0.0935, -0.1337, -0.0010, +0.1682, +0.0380, +0.0001, -0.0007, -0.0007,
-	+0.0006, -0.0009, +0.0017, +0.2981, -0.0003, -0.0741, -0.0007, +0.0668,
-	+0.0477, -0.1859, +0.1886, -0.0001, -0.1307, -0.1552, -0.2429, -0.1578,
-	+0.0800, +0.0001, -0.0007, +0.0020, -0.0491, -0.1470, +0.0004, +0.1592,
-	-0.0737, +0.0672, +0.0001, -0.1061, -0.0550, +0.2227, +0.1435, +0.1894,
-	+0.1484, -0.0763, +0.0967, +0.0308, -0.2037, -0.0901, -0.2943, +0.0990,
-	//18
-	-0.0305, +0.1165, +0.1287, -0.2020, -0.1904, +0.1685, -0.1301, -0.0196,
-	-0.2848, -0.0744, -0.1650, +0.1925, +0.0041, +0.0921, +0.1567, +0.0997,
-	-0.1642, -0.0779, -0.0001, +0.1648, +0.0577, -0.0004, +0.0014, +0.0008,
-	+0.0003, +0.0002, -0.0003, +0.2420, -0.0001, -0.1356, -0.0009, -0.0741,
-	+0.0834, -0.1532, -0.1952, +0.0015, -0.0295, +0.0019, +0.0571, -0.2293,
-	+0.1462, +0.0013, +0.0001, -0.0004, -0.2460, -0.1658, +0.0012, +0.0758,
-	-0.0808, +0.0629, -0.0006, -0.1561, -0.0793, +0.2295, +0.1258, +0.1096,
-	+0.1092, -0.2464, +0.0950, +0.0181, +0.1049, +0.0975, -0.0335, +0.0917,
-	//19
-	+0.0351, -0.0122, +0.2250, +0.0614, +0.0614, +0.2307, -0.3622, +0.1324,
-	+0.1173, +0.0616, -0.0945, +0.0484, +0.0692, +0.0053, -0.0336, +0.1762,
-	-0.1893, -0.0028, -0.0001, +0.0986, +0.0831, -0.0009, +0.0010, +0.0026,
-	-0.0008, -0.0006, -0.0003, +0.0356, +0.0000, -0.0307, +0.0017, -0.0985,
-	+0.0842, -0.2074, -0.0394, -0.0004, -0.1312, +0.0484, -0.2349, -0.0197,
-	+0.1920, -0.0005, +0.0011, +0.0015, -0.0759, -0.0120, -0.0010, +0.0142,
-	-0.1323, -0.0255, +0.0008, -0.1792, -0.0355, +0.2631, +0.0680, +0.1625,
-	+0.1786, -0.2487, +0.1460, +0.1019, +0.2607, -0.0342, -0.0860, +0.1556,
-	//20
-	+0.0264, +0.1371, +0.1032, -0.2805, -0.1354, +0.1126, -0.0835, -0.0218,
-	-0.0698, -0.0619, -0.1355, +0.2303, -0.0087, +0.1087, +0.0497, +0.1169,
-	-0.1251, -0.0931, -0.0002, -0.0043, +0.0417, +0.0009, -0.0008, +0.0007,
-	+0.0006, +0.0002, -0.0010, +0.3026, +0.0012, -0.2144, -0.0010, -0.0502,
-	+0.0470, -0.1658, -0.0541, -0.0020, -0.0828, -0.1087, +0.1445, -0.0476,
-	+0.0066, +0.0006, -0.0021, -0.0007, -0.1326, -0.1629, -0.0008, +0.0736,
-	-0.0579, +0.1449, -0.0001, -0.0517, -0.0653, +0.1763, +0.1323, +0.1857,
-	+0.1333, -0.1172, +0.0961, +0.0162, -0.0781, -0.0466, -0.0294, +0.0736,
-	//21
-	+0.0774, +0.2557, -0.0480, -0.2746, -0.3141, -0.0054, +0.0661, -0.0357,
-	-0.1825, -0.0268, -0.3164, +0.3275, +0.0282, +0.0840, +0.1664, +0.1447,
-	-0.1584, -0.0599, +0.0004, +0.2176, -0.0260, +0.0009, +0.0020, +0.0011,
-	-0.0000, -0.0002, -0.0009, +0.1985, -0.0005, -0.1150, +0.0003, -0.1694,
-	+0.0689, -0.2120, -0.3423, +0.0012, -0.0137, -0.0653, +0.1518, -0.3196,
-	+0.0418, -0.0001, -0.0002, +0.0019, -0.3952, -0.2258, +0.0003, -0.0410,
-	-0.3160, +0.1169, +0.0013, -0.0829, -0.1139, -0.0204, +0.2290, +0.1876,
-	+0.1956, -0.2626, +0.0985, +0.0587, +0.3142, -0.2336, -0.2823, +0.0593,
-	//22
-	+0.4221, +0.1385, -0.0797, -0.2832, -0.4008, -0.0929, +0.0634, +0.0156,
-	+0.0845, -0.3586, -0.3662, +0.5097, +0.2763, -0.1964, +0.2568, +0.1333,
-	+0.2808, +0.2787, -0.0000, +0.3051, +0.2848, +0.0008, +0.0004, +0.0006,
-	+0.0011, +0.0015, -0.0014, +0.0301, +0.0002, -0.2382, -0.0003, -0.6114,
-	-0.1717, -0.0915, -0.6118, -0.0009, -0.0967, -0.4326, +0.4761, -0.2196,
-	+0.0668, -0.0001, -0.0009, +0.0001, -0.0590, -0.0923, +0.0007, +0.5462,
-	+0.1415, -0.0109, -0.0003, -0.2102, -0.0296, -0.2474, +0.4033, +0.9525,
-	+0.0946, -0.7545, +0.3012, +0.0215, +0.4868, -0.1123, -0.3776, -0.0184,
-	//23
-	+0.0730, +0.0596, +0.1569, -0.0901, -0.0518, +0.1582, -0.1851, +0.0329,
-	+0.0579, +0.0308, -0.1131, +0.1100, +0.0233, +0.0054, -0.0016, +0.1529,
-	-0.2043, -0.0451, -0.0006, +0.0638, +0.0664, +0.0019, +0.0009, -0.0004,
-	+0.0011, -0.0002, -0.0003, +0.1299, +0.0016, -0.1112, -0.0004, -0.1295,
-	+0.0784, -0.1499, -0.0963, -0.0009, -0.1004, +0.0174, -0.0687, -0.0415,
-	+0.1291, -0.0004, -0.0015, +0.0007, -0.0909, -0.0732, +0.0005, -0.0273,
-	-0.0847, +0.0521, +0.0008, -0.1283, -0.0438, +0.2376, +0.1066, +0.1484,
-	+0.1610, -0.1803, +0.1262, +0.0632, +0.1948, -0.0479, -0.0991, +0.1049,
-	//24
-	+0.3117, +0.1788, +0.0671, -0.3081, -0.0957, +0.1004, +0.0763, -0.0676,
-	-0.2451, -0.1567, +0.0569, +0.0268, +0.0825, +0.1525, +0.2440, +0.0608,
-	-0.3412, +0.0721, +0.0002, -0.0913, +0.1063, -0.0015, +0.0017, +0.0027,
-	-0.0014, -0.0027, -0.0008, +0.2128, +0.0002, -0.3646, -0.0008, -0.2293,
-	+0.0307, -0.1424, -0.2396, +0.0004, -0.0184, +0.0235, +0.1501, -0.1810,
-	+0.1083, +0.0005, -0.0013, +0.0013, -0.0398, -0.2459, +0.0004, -0.0567,
-	+0.0030, +0.1694, +0.0005, -0.1893, -0.1048, +0.3268, +0.0842, +0.1447,
-	+0.0538, -0.1025, +0.1314, -0.0335, +0.3236, +0.1932, -0.0244, +0.0053,
-	//25
-	+0.4052, +0.4961, +0.1398, -0.0073, -0.4697, +0.1489, -0.4627, -0.3504,
-	-0.1077, -0.0811, -0.6809, +0.2022, +0.0269, +0.3777, +0.4814, +0.1048,
-	-0.1351, -0.0812, +0.0015, -0.0633, -0.0287, +0.0004, -0.0012, +0.0015,
-	-0.0006, +0.0008, +0.0010, +0.0157, -0.0007, -0.1289, -0.0005, -0.3270,
-	+0.0113, -0.2310, -0.2660, +0.0017, -0.1703, -0.7104, +0.4676, -0.2159,
-	+0.2716, +0.0003, -0.0005, +0.0005, -0.3787, -0.0050, -0.0004, +0.3120,
-	+0.2226, +0.0493, +0.0013, +0.0395, -0.1446, -0.1229, +0.1919, +0.4883,
-	-0.0575, -0.2365, -0.0734, +0.2716, +0.6544, -0.2428, -0.4467, +0.1654,
-	//26
-	-0.2479, +0.0615, +0.1741, +0.0162, -0.5825, +0.0594, -0.7230, -0.0260,
-	-0.0485, +0.0896, +0.0672, -0.1092, -0.0405, +0.1098, +0.0152, +0.1011,
-	-0.3244, -0.2129, +0.0000, -0.0393, +0.1475, -0.0001, -0.0019, +0.0029,
-	+0.0002, +0.0017, -0.0006, -0.1210, +0.0002, -0.2875, +0.0007, -0.4754,
-	+0.0530, -0.1766, -0.3640, +0.0001, -0.1330, +0.2328, +0.1096, -0.0428,
-	+0.1773, +0.0027, -0.0006, -0.0013, -0.2937, -0.0576, -0.0009, -0.0198,
-	-0.6996, +0.1235, +0.0020, -0.0545, -0.0644, +0.2087, -0.1430, -0.3335,
-	+0.0730, +0.0991, +0.0946, +0.0446, +0.4431, -0.1680, +0.0230, +0.0083,
-	//27
-	+0.0000, -0.0025, +0.0000, +0.0000, +0.0014, -0.0004, +0.0003, +0.0002,
-	-0.0010, -0.0018, +0.0002, -0.0008, +0.0007, +0.0011, +0.0011, -0.0031,
-	-0.0006, +0.0005, +0.0026, +0.0012, -0.0018, +0.0015, -0.0004, +0.0003,
-	-0.0009, -0.0005, +0.0005, +0.0006, +0.0008, -0.0007, +0.0008, -0.0008,
-	-0.0019, +0.0027, +0.0005, +0.0017, +0.0010, +0.0003, +0.0011, +0.0009,
-	-0.0009, -0.0009, -0.0004, -0.0011, +0.0005, +0.0011, -0.0008, +0.0010,
-	-0.0001, -0.0017, -0.0007, -0.0007, -0.0010, -0.0027, -0.0006, +0.0001,
-	-0.0009, -0.0019, +0.0000, -0.0001, -0.0007, +0.0013, -0.0003, -0.0013,
-	//28
-	+0.2008, +0.2329, +0.0325, -0.3651, -0.4689, -0.2651, -0.3326, -0.0647,
-	-0.7383, -0.1876, -0.2019, +0.2806, +0.1836, -0.0012, -0.2708, +0.2003,
-	-0.2847, -0.1943, -0.0008, +0.0051, +0.2133, +0.0006, -0.0005, +0.0011,
-	-0.0002, +0.0010, -0.0004, +0.1743, +0.0000, -0.2272, +0.0015, -0.0601,
-	-0.0447, -0.0816, +0.1596, -0.0011, -0.2204, -0.2789, +0.2830, -0.0233,
-	+0.0361, +0.0006, +0.0012, -0.0009, -0.0908, -0.0920, +0.0001, +0.1132,
-	-0.2440, +0.2044, -0.0013, +0.0043, -0.0162, +0.5016, +0.5223, +0.1193,
-	+0.0764, -0.1246, +0.1602, +0.1489, -0.0666, -0.3342, +0.4498, -0.1661,
-	//29
-	-0.2866, +0.1061, +0.1953, -0.0187, -0.3525, +0.1779, -0.6952, -0.0381,
-	-0.2794, -0.0858, +0.0091, -0.1148, +0.0437, +0.2320, +0.1462, +0.0317,
-	-0.1127, -0.1511, +0.0000, +0.0525, +0.2125, -0.0002, +0.0000, +0.0031,
-	+0.0004, -0.0004, +0.0002, +0.2340, +0.0011, -0.3199, +0.0013, -0.0735,
-	+0.0073, -0.3315, -0.0867, -0.0005, -0.1468, +0.0255, +0.0563, -0.2035,
-	+0.1641, -0.0011, +0.0009, -0.0027, -0.3826, -0.1024, +0.0001, +0.1450,
-	-0.5216, +0.1925, +0.0006, -0.0217, -0.1108, +0.2139, +0.0312, -0.1823,
-	+0.0319, -0.0110, +0.1178, -0.0005, -0.0630, +0.0348, -0.0523, +0.0454,
-	//30
-	+0.0006, +0.0177, -0.1723, +0.0037, +0.0016, -0.1006, +0.0211, +0.0208,
-	+0.1460, -0.0425, +0.0138, +0.0077, +0.0419, +0.0169, +0.0021, -0.0668,
-	+0.0205, +0.0291, -0.0001, +0.0005, -0.0177, -0.0005, +0.0011, -0.0008,
-	-0.0001, +0.0009, -0.0005, -0.0038, +0.0004, +0.0141, +0.0007, -0.0067,
-	-0.0209, +0.0266, -0.0011, -0.0008, +0.0283, +0.0092, +0.0058, +0.0217,
-	-0.0145, -0.0000, -0.0004, -0.0012, +0.0606, +0.0203, -0.0012, -0.0487,
-	+0.0238, -0.0235, +0.0017, +0.0220, +0.0003, -0.1910, +0.0651, -0.0252,
-	-0.0456, +0.0305, -0.0356, +0.0190, +0.0250, +0.0029, +0.0131, -0.0177,
-	//31
-	+0.5723, +0.4493, +0.0335, -0.0522, -0.1651, +0.0951, -0.2535, -0.3173,
-	+0.0209, +0.0163, -0.0931, +0.1957, -0.3959, +0.2692, +0.3318, +0.1118,
-	-0.2671, -0.0564, -0.0015, +0.2937, +0.1793, -0.0010, +0.0010, +0.0036,
-	+0.0007, +0.0009, +0.0003, +0.3479, -0.0004, -0.0808, +0.0004, -0.0114,
-	-0.0033, -0.0882, -0.3248, -0.0012, -0.2739, -0.1155, +0.1720, -0.0041,
-	+0.4134, +0.0004, +0.0015, +0.0014, -0.5571, -0.2881, -0.0024, +0.1794,
-	-0.2006, +0.3172, +0.0008, -0.0707, -0.1372, -0.1097, +0.0587, +0.2403,
-	+0.1825, +0.1102, +0.2434, -0.0415, +0.1996, -0.1519, -0.2734, +0.1377,
-	//32
-	+0.2235, -0.0089, -0.1719, -0.4333, +0.2023, +0.0432, -0.3600, -0.0067,
-	+0.1826, -0.1181, -0.0741, +0.0324, +0.0516, +0.0856, +0.4092, -0.0786,
-	-0.2289, -0.2439, +0.0001, +0.6469, -0.0361, -0.0015, -0.0011, +0.0033,
-	+0.0018, -0.0000, -0.0017, +0.4013, -0.0007, -0.2810, -0.0010, +0.0196,
-	+0.0931, -0.5306, +0.1758, -0.0000, -0.2092, -0.6634, -0.1485, -0.0660,
-	-0.0349, -0.0000, +0.0004, -0.0013, -0.1061, -0.5817, -0.0007, +0.4457,
-	-0.5122, +0.2464, +0.0011, -0.2380, -0.2329, -0.0766, +0.3303, +0.2732,
-	+0.0942, -0.1698, +0.0170, -0.1066, +0.1109, -0.0894, -0.3529, +0.0655,
-};
-
-#endif //__CONVOLUTION2_H_INCLUDED
