+++ convdata.h
