--- opencv.sh
@@ -1,50 +1,0 @@
-#!/bin/bash
-
-# Program:
-#	Compile the OpenCV project
-# Arguments:
-#   $1:The source file name
-# History:
-#	WangShu  Tue 23 Jun, 2015
-
-if [ -f $1.cpp ];
-then
-	g++ `pkg-config --cflags opencv` -c $1.cpp
-else
-	echo -e "  \033[0;31;1m[Error]\033[0m $1.cpp does not exited!"
-	exit 1
-fi
-
-if [ $? -eq 0 ];
-then
-	if [ -f $1.o ];
-	then
-		echo -e "  \033[0;32;1m[Compile]\033[0m Complete -- $1.o has created!"
-		#g++ `pkg-config --libs opencv` -o $1 $1.o
-	else
-		echo -e "  \033[0;31;1m[Error]\033[0m $1.o has not created!"
-		exit 2
-	fi
-else
-	echo -e "  \033[0;31;1m[Error]\033[0m Compile Failed!"
-	exit 3
-fi
-
-g++ `pkg-config --libs opencv` -o $1 $1.o
-
-if [ $? -eq 0 ];
-then
-	if [ -f $1 ];
-	then
-		echo -e "  \033[0;32;1m[Linking]\033[0m Complete -- $1 has created!"
-		#./$1
-	else
-		echo -e "  \033[0;31;1m[Error]\033[0m $1 has not created!"
-		exit 4
-	fi
-else
-	echo -e "  \033[0;31;1m[Error]\033[0m Linking Failed!"
-	exit 5
-fi
-
-exit 0
